[roll] Roll fuchsia [pbm] Prevent duplicate product bundle downloads
This CL adds a check to the pbms lib to prevent downloading a product
bundle that's already contained in the cache. In the case of invalid or
corrupt product bundles, we add a --force flag to the product bundle get
command to bypass the duplication check and download anyway.

There's also some error-reporting cleanup in the product-bundle tool.

Original-Bug: 98281
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735742
Original-Revision: 4e8708d87602be7997bbc263506ba7ec929e2312
GitOrigin-RevId: f66c57b6b06942d8acf51f81402b1707f4b5b065
Change-Id: Ida31598133896ac7e979e8a1375dec8f53b5b734