[roll] Roll fuchsia [pbm] Clear cached metadata when removing bundles
The attached bug describes a use case where the user was confused by the
contents of the pbm cache after calling `ffx product-bundle remove`. The
root issue is that the metadata (product_bundles.json, etc.) is not
removed by the tool even if all of the bundles are, so they still show
up in a listing with --cached.

This CL adds logic to the remove command to clear the metadata for a
bundle if that bundle is the last one to be removed from a subdirectory.
In other words: since PBv1 bundles share a metadata file and directory,
we only remove those if all of the bundles it applies to are also
removed.

Original-Bug: 111641
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737548
Original-Revision: 84375c6d5204c7107fe7115e73b6b9f8bac248c7
GitOrigin-RevId: 7c4c241db54939b8189993892a9ee116f6cb6a79
Change-Id: I885789ed5e47a970bbe6eb5893e321bc211bd23d