[roll] Roll fuchsia [fidlc][docs] Document ErrResponsesWithErrorsMustNotBeEmpty
fi-0076 and fi-0077 are a pair of error messages that explain the
different syntax required for response types when it should hold no
information. This is a bit inconsistent at the syntactic level but this
CL documents the current state.

Test: fx test fidlc-test; manual devsite staging
Original-Bug: 111278
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738006
Original-Revision: 9f089ad2264786942e4fc3177b18fca28b2bc471
GitOrigin-RevId: b6151674af46701b85116a06da410e318d66fb31
Change-Id: I9c46c54abc8af6cd707e91d18d47b2e41099b5ea