[roll] Roll fuchsia [netstack] Add ctx to interface watcher event loop
Add a context.Context to the interface watcher event loop so that it
can be cancelled.

Add a unit test helper for starting an event loop and deferring a
function for waiting for the event loop goroutine to complete to avoid
leaking the goroutine and tripping the checker.

Test: fx test netstack-gotest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734223
Original-Revision: 97dbb3d1fd024f0218733abfa2378cb8caa8feb1
GitOrigin-RevId: 35749231b3e731c719b1ae1e3ec228e4ab191358
Change-Id: I594630bab8aa7e34be468ae6acce588a00a1f93f