[roll] Roll fuchsia [fidlc] Merge protocol content errors.
Merge ErrUnrecognizedProtocolMember and ErrExpectedProtocolMember into a
single ErrInvalidProtocolMember.

There was not really a meaningful distinction between these two error
cases in the compiler, as both could occur in cases where a protocol
contained something other than a valid protocol member. Consider the
following examples.

protocol Example {
  NotAMethod;
};

protocol Example2 {
  42: TableLikeMethodDecl();
};

In Example1, NotAMethod receives ErrUnrecognizedProtocolMember while in
Example2 TableLikeMethodDecl receives ErrExpectedProtocolMember. Both of
these are situations where a protocol declaration should contain a
protocol member but contains something which doesn't actually match the
format of a protocol member. There's not really a meaningful distinction
between the two cases that can be described in the error documentation,
as both are 'some member is invalid, change to a valid protocol member'.

The one edge case is the case of a completely empty protocol member,
that is a plain semicolon:

protocol Example3 {
  ;
};

This received ErrExpectedProtocolMember, which may make slightly more
sense than ErrInvalidProtocolMember in that the semicolon imples that a
member should exist but doesn't, but I think it is reasonable to
classify "empty statement" as an invalid member, so I don't think this
should block merging these two errors.

Original-Bug: 111278
Test: fx test fidlc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738704
Original-Revision: 23607df934b64a478ff1943d07a3906179317dd2
GitOrigin-RevId: 29673a6f0c867c8032d49e1453264c74ccda6940
Change-Id: I05b0b6399a44c6a21eafb38f56ce78b3d5c8f4db