[roll] Roll fuchsia [flatland] Fix flake in flatland_manager_unittest
A couple of tests rely on a couple of dispatcher hops to complete but
weren't correctly waiting for that to happen. This CL fixes that.
Specifically:
1. We must wait for all the destructors of all Flatland instances to run
(which does not happen on the main thread).
2. We must then run the main thread dispatcher until empty to ensure the
calls to FlatlandManager and FlatlandPresenter finish up.

Original-Fixed: 110682
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738004
Original-Revision: 00f26c9ee1d5dd96527988339d1b320da0927cce
GitOrigin-RevId: fe6eb858db4cb08c6b80e0d6948e286b558ce2aa
Change-Id: Ie6f82c76f713c861ceaa0540ceabe603fd08d8a4