[roll] Roll fuchsia [zxcrypt] Use DdkChildPreRelease to watch for removal
This is the officially-supported way to remove a child device and be
notified by device_manager when the device is about to be released.  We
can use this instead of rebinding the parent device, which will also
give us the ability to avoid replying to the `Seal()` call until
`device_manager` has also fully unbound the child, which means that a
subsequent `Unseal()` will not attempt to DdkAdd a child device with an
existing name.

Original-Fixed: 111714
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737988
Original-Revision: 8cd1197515189cc2dba2547740f7626433669e1f
GitOrigin-RevId: 459c197c2a27b34f41ff7d6b045d1a48e4104493
Change-Id: I24a7ec9c6cbbfa5c8ee1b9757e493b0bf5d888b4