[roll] Roll fuchsia [docs] Document standard storage capability names.
This CL follows on from earlier conversations in the API and CF chat
rooms to define the standard names and behaviors we assign to the
commonly used storage capabilities.

These storage capabilities form a part of the API between a component
and the parent supplying its capabilities so Fuchsia.dev isn't the ideal
formal home for these definitions in the long term, but its better than
nothing.

Documenting these is a small but necessary step along the road to
agreeing a stash replacement so I'm tagging this with that bug.

Original-Bug: 102879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734146
Original-Revision: 28ec57164d71396d29d4173bf6ef5fd67dfd2ce7
GitOrigin-RevId: c05a8825b6c13efc60035161e0790b98dc8f0565
Change-Id: I7e3e151b03b663a6addce0ef95ba7d0a1f7b0419