[roll] Roll fuchsia [zx][kazoo] Enforce proper FIDL naming in library zx
zx features inconsistent naming, much of which is invalid in the eyes of
the FIDL style guide. The library does not need to be a tall poppy in
this respect, and so this change
* updates the naming of almost every* declaration to be conventional, and
* enables the build-time enforcement of naming remaining conventional
going forward.

"Almost every" as there are a handful of snake case names exported in
the SDK in zx_common.fidl, some of which used prolifically (e.g.,
`zx.rights`) and more care will be needed to migrate these. For these
cases, there are already exemptions in the linter to ignore them.

Original-Bug: 109734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725184
Original-Revision: ee6318008eaeaaf483c2f9a8870fcf92dcf5a68d
GitOrigin-RevId: 69a4b7d19de30c604548a17a8d9b6b6a85ecfaa9
Change-Id: Ia040ceb4ee212ebbc789d915dce66fa5a86ea508