[roll] Roll fuchsia [emu] Add deprecation warning and pause to `fx emu`
The `fx emu` tool will be retired at the end of October. This CL changes
the info banner about "early adoption" to a deprecation warning, and
increases the pause to a full 5 seconds. This gives the user more time
to observe the message, and encourages them to transition.

Original-Bug: 98983
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738403
Original-Revision: 67b01eeea29b01a740edd48f16147b293cf0fa9a
GitOrigin-RevId: 611d9d4bc3987693d7412614139065406a1b0d43
Change-Id: Ib644fa1564aaa2a6d2e329cf4438f8f78c9adc13