[roll] Roll fuchsia [fdf] Pass drivers_serial_log test in DFv2
Pass //src/tests/drivers_serial_log:bringup_tests test with DFv2
enabled. The purpose of this test is to ensure the driver framework can
print logs to serial. In order to do this, the test waits for a certain
log statement to be printed that should be sent from the driver manager.
This log statement is not printed when DFv2 is enabled and so this test
must be updated to wait for a log statement from the driver manager
found in both DFv1 and DFv2.

Bug 106650

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736122
Original-Revision: 629c4c104ec8cd3482da19c5f19ca0349fb6e3a5
GitOrigin-RevId: 9d5e13acc7ac0ac912c598d7b0580734323588b4
Change-Id: Ic84770a581af4890674d54504eb22808b56248f5