[roll] Roll fuchsia [fuchsia-tpm-protocol] Rename fuchsia-tpm fuchsia-tpm-protocol.
This name is more explicit as to what this library implements
it implements the fuchsia.tpm protocol and isn't the "fuchsia-tpm"
library. I will be using the fuchsia-tpm library in a future change
to refer to the generic fuchsia-tpm library usable by components.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737603
Original-Revision: 769cfc89020f18af45d850a2d721f4504a032182
GitOrigin-RevId: 75cd4360cc789dbbe5b5bcf53dafbce3960ebcb9
Change-Id: I2789fc1d88105c45c886958350a354701d0b851f