[roll] Roll fuchsia [system-tests] Also retry IsDeviceUpToDate downgradeOTAAttempts times
The target may reboot during the "what version is on the target?"
pre-check as well as while we're trying to run the OTA, so try to be
resilient to that as well.  Using the same config flag makes sense here:
both are aiming to handle the downgrade build spontaneously rebooting.

Original-Fixed: 110451
Original-Bug: 109499
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738445
Original-Revision: cda1c907e5d5ab16e21aeb94272e9e553e9fb0e4
GitOrigin-RevId: be9eac9aa8a34531f17e6d54f524b2cc491aea85
Change-Id: I74495e1e3a751b7a0bdbf71b2b279c53ae6c91d9