[roll] Roll fuchsia [fidlc] Avoid duplicate errors for @available in unversioned library
This CL makes the AvailabilityStep mark an Avaibility as kFailed if it
bailed out because of ErrMissingLibraryAvailability. Previously it
didn't, so on visiting its children it would recursively try to visit
the parent again and hit the same error, making a duplicate message.

Test: fx test fidlc-test -- -f "VersioningTests.BadUseInUnversionedLibraryReportedOncePerAttribute"
Original-Bug: 106549
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738443
Original-Revision: 714fdd7d7aeac42823d1e3cb504746476eb86a76
GitOrigin-RevId: e22b537c3af5ae7cc18726f0028e9a2b40e3cc14
Change-Id: Id64ccc4dfe06c3ef8389573240db3457a047e474