[roll] Roll fuchsia [netstack3] Remove rcv_residual for post FIN receipt states
This was added due to a concern for Core to be platform agnostic - For
example on a platform where netstack lives in the kernel, and we need to
keep all the bytes locally in stack, then we need such field to keep the
bytes that has been received but not read by the user yet for these
states. This is not a concern for Fuchsia at all and to support such a
platform doesn't seem to be near term. Removing it as a recent attempt
to remove the #[allow(unused)] lint reveals it is not being used. We can
add it back when the need arises.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739648
Original-Revision: f58a1711d0fffc9630a6d8b55a145f970ac639cb
GitOrigin-RevId: 642f3e615d6f51e48d64b320dfb054bbcf1784bd
Change-Id: Idf4ce55c8e27c1f4c8708c4774fa421c8133753d