[roll] Roll fuchsia [aml_mipicsi] Only join thread if created
thrd_t doesn't have a sentinal null value and if AmlMipiDevice::Create
fails, the thread is attempted to be joined with an invalid value
leading to a crash. Wrapping it in a std::optional allows us to know
whether or not it was initialized.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738645
Original-Revision: a4f9f1f543caa69d68b92092843a833a1623aa79
GitOrigin-RevId: c18973ca17cc017221bf09aaf68a377204622969
Change-Id: I1a3a00550b145bb5215aae93c3cc89da3e8bc85d