[roll] Roll fuchsia [musl] Remove notice
This CL removes the custom copyright notice from this file, which makes
it easier for our customers to do their compliance. Without a custom
notice, this file can be treated the same way as all the other files in
this directory, which should be accurate from a compliance
point-of-view.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739577
Original-Revision: 7892075151ef489a73a9135e37da81fbaeff0dec
GitOrigin-RevId: ab34a7c4e1b88084a88baf3fa26dab3d6aa85ac6
Change-Id: Ia3a9cba8a5ecb6fcbc0d921a0380200bd455a5c5