[roll] Roll fuchsia [pbm] Reverse isatty conditions for output
The attached bug describes a bug in the product-bundle output, where the
`get` command is failing to print progress to the terminal. This is
because the check for a TTY output was reversed, so it would only print
on non-interactive outputs instead.

This CL switches the condition, so it will skip output only for
non-interactive sessions.

Original-Bug: 111861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739563
Original-Revision: 71d57a3c59b13c579e4e040921ccf5f4c52b64f4
GitOrigin-RevId: ae93e86f22083236eef0f21c1953a36bb41fc102
Change-Id: I3aba67b83f1051e24a68148b01df6d2668b09343