[roll] Roll fuchsia [pbm] Reordering PBM messages on duplicates
The attached bug points out that when attempting to `get` a duplicate
PBM, the tool was printing the duplicate repo message instead of the
duplicate bundle message. This was because the sequence of the checks
was reversed.

This CL puts an additional check for duplicate bundles before the check
for repos, so it's clear why the tool doesn't fetch the data. It will
still print the repo message if the user attempts to download a bundle
with the same name from a different version, as expected.

Original-Bug: 111862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739266
Original-Revision: 4129b92325ba59589b38d98d2e414d6527467856
GitOrigin-RevId: 192921292867f9def362160cb56ef8a9e2630170
Change-Id: I84d3fb54812be158e13346e863d73541d8a17781