[roll] Roll fuchsia [pbms] RepoClient must update before refreshing
Originally RepoClient::new would auto-update the metadata to the latest
version before, but this was removed in a refactor. Unfortunately we
missed fixing this in the pbms library because there wasn't a test for
this functionality.

This patch moves the repository functionality into a dedicated module,
and adds the correct call to `RepoClient::update`. It also adds a test
to make sure an issue like this won't be missed in the future. We
unfortunately can't mock gcs, but this code was refactored to make it
mostly not dependent on the presence of gcs.

Original-Fixed: 111508
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737854
Original-Revision: d159d3d0b73eb8d9d614aea2bf76d724a9c393b4
GitOrigin-RevId: ff7b0d178759ac449f17de1657134b557385a4c0
Change-Id: Ie3950faaee6448d8324494b93262c4a8260f64d5