[roll] Roll fuchsia [pbm] Reverse sort order of product-bundle list
The attached bug points out how the output of `ffx product-bundle list`
sorts the bundles in alphanumeric order, which is the opposite of the
rest of the product-bundle commands and puts the newest bundles at the
bottom of the list.

This CL adds a call to reverse the order of the list, such that is
matches the rest of the product-bundle commands and puts the newest
bundles first.

Original-Bug: 99226
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739567
Original-Revision: 05c94682cffe9a96001bdfc8b1c9f4908a16a093
GitOrigin-RevId: e32ce99f7824675b3da7d35cd236cb9acdb3610c
Change-Id: Ib5c3b269166912194ad4381d5aacc0c2b45a2146