[roll] Roll fuchsia [ffx] Remove ffx buildid from config.
This is a value that has no need of being looked up in the config. It is
currently used to determine whether or not the daemon is the same binary
as the one invoking it, and all of this is something that can be looked
up inside the binary itself.

Without this change, if ffx were to force a restart it would send its
runtime config, which gets really confusing if there are actually
different versions of ffx at play, as this buildid in the config would
not match the one of the actual binary.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739654
Original-Revision: 67fd34192645540713c3cae73359b045aa627149
GitOrigin-RevId: c3d2f3d5215caf3d2ab7e22dffcee42d554ea633
Change-Id: Ic44ee264fd0549d8c1718e6443fcec98b88a877f