[roll] Roll fuchsia [pb] Add a transfer_manifest crate for upload/download of artifacts
This crate is reponsible for declaring the schema for the transfer
manifest, which is used to declare the local and remote paths for a set
of artifacts. This manifest will be used to upload all the artifacts for
a product bundle.

Original-Bug: 111660
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737862
Original-Revision: 0af7b1543866790f8329be37a1d5fb38643d5334
GitOrigin-RevId: 76b0da972ea6d979fb968e25a167daa0a6b7083d
Change-Id: Ia4fe4c5bf2027bd5ab209ea03378232d46cbf80a