[roll] Roll fuchsia [settings] Add unit test to verify light controller calls without restore
In fxr/715324, an EnvironmentBuilder test was added to verify that Set
calls and the OnMicMute event don't panic when they're made after the
light controller times out and is started again. That test was a quick
test we added as the fix needed to go in quickly, so it indirectly tests
the behavior at the service-level and has to wait 6 seconds per test for
the controller to timeout.

This CL moves those tests over to a unit test that creates the
controller and calls methods directly, which no longer depends on the
controller discarding behavior.

The test itself is still pretty verbose and a good example of a test
that would benefit from mocking out objects.

Original-Fixed: 107545
Original-Bug: 103274
Test: setui_service_tests
Multiply: setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738902
Original-Revision: 01a731887b7ee0abc461997f260f70d284df4842
GitOrigin-RevId: f0ba2d84b73d595c6075edad092eee67c4084365
Change-Id: I8182aa5522e127728291c43ec9fda755fd05a837