[roll] Roll fuchsia [netstack3] Remove #[allow(unused)] attribute on TCP mod
This attribute was added to suppress errors at an early stage when not
much from the TCP module was referenced by Bindings. Now that we are
actually using it, we can remove this and fix mistakes that have slipped
through while this attribute was on.

Original-Fixed: 95688
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739443
Original-Revision: 731b6b7a098644661b0ca8edcd66c9b7190da8cb
GitOrigin-RevId: 9dc25f9b6f7c2c15932ed10f37b3518f6a3f9511
Change-Id: I5874a287280909503e4454eb1d800b68ad1815e9