[roll] Roll fuchsia [credential manager] Log error for the synchonization failure.
With the problem in fxb/111926 I was running into this. Prior to this CL
if the disk and Cr50 get out of sync credential manager just silently
dies without logging anything. This doesn't fix any of the root causes
but at least if we're looking at logs where people were in a
missynchronized state we'll know it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739780
Original-Revision: eb67dcca20307dc610e7fed72bfd772c71e6eec5
GitOrigin-RevId: 5d9bdeca1bee9d069ab9d62541a8cf07a8d46303
Change-Id: I738246e24e98441d0ae393edfe399c2f297534a9