[roll] Roll fuchsia [netstack3] Back `DeviceId` with references
This change introduces a weak pointer to device IDs so that device state
may be accessed without needing to lookup an ID in a devices map which
requires taking a lock (introduces extra contention) lock and cloning
a strong reference.

Note that this change still incurs the cost of incrementing the strong
reference count since device IDs are backed by weak references which
need to be upgraded to strong references to access the pointed-to
state.

Also note that `EthernetDeviceId`s still keep the scalar (`usize`)
value to act as a human-readable ID. Values like 0, 1, 2 are much
easier to remember/follow instead of pointer values like 0x990184012
and 0x991094012 when reading logs.

Original-Bug: 48578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738647
Original-Revision: 87e1fdae7b1b95ccee7b59056142e338fe70cd99
GitOrigin-RevId: b8c3441a41c65b3a81556d85865ffdced8dbd3a3
Change-Id: I897bd1a3ef3483269a442c2ed5ebe647576cfbe3