[roll] Roll fuchsia [docgen] Format links in comments.
Looks for markdown links with no destination and attempts to look up the
symbol name and automatically create the proper link destination for it.

Move the existing comment formatting code to a new file since it's
getting more complexity.

Enhance the documentation. Remove the obsolete instructions on running
separately and document how docstrings and titles work, along with the
new link behavior.

Make the _toc writing code order stable. Previously this was in map
order which is based on hash functions.

The current code does not handle non-toplevel symbols (e.g. with classes
and namespaces, or anything with a "::" in it). This will be addressed
in a followup.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739444
Original-Revision: c7a3c781a744d73843fc43f3648b63dd64036f00
GitOrigin-RevId: a514f9480326c2ec1d7a44f1b19fd49a4d9e8561
Change-Id: I1950f1692509c62f3b3e92d4d9ff01e576f233c0