[roll] Roll fuchsia [kernel][x86] make sure top level page table is <4GB when bootstrapping secondary cpus
In specific situations (KASAN on machines with low <4GB memory) by the
time the bootstrap aspace is created all of low memory may be exhausted.
This causes problems when bootstrapping secondary cpus as they go
through the real mode -> 32bit protected mode -> 64bit protected mode
bootstrap.

Make a copy of the top level page table for the bootstrap aspace in low
memory just prior to starting the secondary cores to make sure this
constraint is met. Since the aspace is static at this point and this
copy of the top level page is only used as the cpus bounce through the
modes it is safe to use a copy.

Original-Bug: 96565
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739884
Original-Revision: 772df64c5ae890919d004a3aa3b1bd5036417a8e
GitOrigin-RevId: 546c1e686622c7827e72486cf44a4477df8ea2c5
Change-Id: I9e6a3d180092bc0ce9f31dc768793a3816582cb0