[roll] Roll fuchsia [check-licenses] Build API contract updates.
Move the build_api_module target out of the check-licenses directory.
Other projects define theirs in the root //BUILD.gn file, so I will do
the same in a future CL.

Also:
Previously, license files and segments were stored in a subdirectory of
the "license" folder, in the output directory. By default it still is,
but when uploading these files to GCS, it's probably better to remove
paths / details that are implementation-specific to check-licenses.

This CL adds an optional path argument, when provided it will save the
license files to that path instead of the default one.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740369
Original-Revision: 6390d54d16f211169d3efd719be8fbf4c1c34959
GitOrigin-RevId: 452754c23856aebc0f456202a0f23ef3e845a8a5
Change-Id: Iab24ddf78f0a4b2050e8d15529bcbca0860a96ed