[roll] Roll fuchsia [zx][kazoo] Introduce @inout
This CL introduces the `@inout` attribute for syscall parameters, which
signifies that although they might appear in the input 'request' struct
in this representation, they are intended to be mutable 'in-out'
parameters in the syscall interface proper.

The attribute replaces the use of the magic "Mutable"-prefixed aliases
that effectively signified the same thing. Ultimately, @inout is an
expedient placeholder until we are able to model things better (or at
the very least until we have deprecated kazoo).

Original-Bug: 110294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726630
Original-Revision: 729a9874aa26918366fb951bfc94c3285eedf4c0
GitOrigin-RevId: 50d0d33ffd3ef3da2cb79fd2cbfc121beef8b293
Change-Id: I573e511daed54df0821d9196ae8226ebd0be348e