[roll] Roll fuchsia [build][bazel] Enforce user_output_dir in build directory.
Enforce the Bazel output_user_root to be a directory under
the build directory (gen/build/bazel/output_user_root).

The default is to use a system-specific user directory, but
this creates issues on infra when USER is not defined in the
environment.

Original-Fixed: 112206
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741724
Original-Revision: e832589374c37da233ce30c0ef6848d70ae1f672
GitOrigin-RevId: 0fc22fbbbed8a4bf0b9cbb80169b7f018283894d
Change-Id: I51b13c6f379dac40a645b53c88056334a7ef4196