[roll] Roll fuchsia [component][fuzzing] Add timeout to Fuzzer::stop
This CL adds a timeout to the method that implements
`fuchsia.fuzzer.Manager/Stop`. If a fuzzer is hung, e.g. stuck in a
blocking read, it may not be able to handle the closure of the
ControllerProvider channel from the registry and exit.

Without this change, the manager will be stuck waiting for the
fuzzer's output to finish, and the client to the manager will be
stuck waiting for a response to the call to `Stop`.

With this change, the manager will notice the fuzzer's failure to
exit and ask the test_manager to kill it, ensuring an eventual response
to the client.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741523
Original-Revision: 6dff89212d2c04835d599cfddddd2ed3eafe9db8
GitOrigin-RevId: 3595cb06cdd62eab3076e2b8b73a20607954e267
Change-Id: If68b19b19b50ffd2d3b2575c6d57dfdc976480bf