[roll] Roll fuchsia [fshost] respect fvm_ramdisk flag
When the fvm_ramdisk flag is provided, blob and data are required to be
on a ramdisk. Any fvm partitions not on a ramdisk will still be bound,
and zxcrypt will be bound if appropriate for the data format, but
filesystems on them won't be launched.

Original-Bug: 87511
Original-Fixed: 109334
Test: fx test fshost_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739529
Original-Revision: 155f0b5ee7f2f93facdd3901da6e349a11f5aa24
GitOrigin-RevId: e5d16e71c70756b00df82ea7362521d8706184c5
Change-Id: Ieaa0d763f0670c57b55b48fe11cacabdcec7456d