[roll] Roll fuchsia [swd] Generate new devhost roots to remove keyid_hash_algorithms
`keyid_hash_algorithms` is a nonstandard extension to TUF metadata,
which was in place when we generated 7.root.json. rust-tuf warns when we
see metadata signed with unknown key ids, so when we switch over to
`package-tool repository publish` the build will always emit this
warning.

To avoid that, this generates 8.root.json and 9.root.json to allow us to
transition to metadata that no longer references this field and avoid
the warning.

Original-Fixed: 111783
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739776
Original-Revision: ec41d152ef8a092d87643dc8c3bece7780ef5f8d
GitOrigin-RevId: cf7f17c30eecf0a0aa3036d7d4194e39541585e7
Change-Id: Ia6490c85147eb810adf70b9e971fb3ceaa039efd