[roll] Roll fuchsia [debugdata-item] Add kernel-zbi-test for debugdata
Add a test that verifies that arbitrary debugdata items are plumbed
all the way through to userspace and are correct.

The test uses the turducken cuckoo zbi to use a test that appends
two debugdata items, and then boots into zircon for a userspace test
to check that the contests are exactly what is to be expected.
Contents, order and quantity both verified.

Test: fx run-zbi-test early-boot-debug-data-item-zbi-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727793
Original-Revision: 16de5f320a365aca95563aa556c78a1270c30faa
GitOrigin-RevId: cb4f8400887a5361e6726a231ef73377df7afbc9
Change-Id: Ic75275b881028ab411265e9d79c393fa62806ce0