[roll] Roll fuchsia [fxfs] Prevent truncate from exceeding transaction size
This splits up truncation so that it is performed over multiple
transactions. If we do that, then we need to handle the case where not
all of the transactions get replayed. To address that, we can reuse the
graveyard: if we need to use more than one transaction, we insert a
graveyard entry with the first transaction and then remove it for the
last entry.

Original-Fixed: 110843
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740082
Original-Revision: 6898b2ff612d514cae701cf19822d8434f1d1c40
GitOrigin-RevId: 36809043182075055c319d81e569318b1fc550ea
Change-Id: I74ef04413e944cdcaa512910d60cee8afaa24a78