[roll] Roll fuchsia [run-test-suite][ffx] Log durations for suite execution
Adds a simple log that records how long it takes to execute a test
suite, after the connection to test_manager is already established.
This will allow us to compare performance between ffx and
run-test-suite, and see if the performance gap is primarily in the
setup steo, or the execution step.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742464
Original-Revision: 9d7af9a1a113a3265d2213bacd072cfd918ca3b4
GitOrigin-RevId: c3ed7bb5abb1714bed96252b1c3ac20da459def4
Change-Id: I3d8d974fb433b2a391ee722468b08a797ea3ed53