[roll] Roll fuchsia [docgen] Improve docgen links.
Create links in the toplevel reference for the library (previously the
linkifier was not called for this text).

Format cross-references as HTML links rather than markdown links. This
allows us to avoid linking the (...) section which may follow the name.
It looks incorrect to link that in cases where parameter values are
specified.

Adds more usage advice to the README.

Adds () to the functions in the index page. This makes them look more
like function calls. Adds struct/class/union tags to those index items
to clarify what they are.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742227
Original-Revision: c1fe4aa96837788b58be7069d0117ca969db31de
GitOrigin-RevId: 45f9afe020c116331ca4f0f07ca91c5ba41c613c
Change-Id: I5921f6f40bd377b474b43082856756d44b00ca18