[roll] Roll fuchsia [amlogic-display] Fix initialization of TV070WSM_FT LCD
Until change Ibf2bda6ceb05f3192125bc03bf5124e760399857 , opcode
0xFD was treated as identical to a write command. With that patch it's
treated as kDsiOpDelay, which is a complicated sleep, so most of the LCD
initialization was being ignored.

To fix this, refresh the init codes for this panel. All the 0xFD entries
are converted to 0x23, which is a short write command.

Original-Bug: b/252224823
Original-Bug: 112367
Tested: "display-tweak panel --power off && display-tweak panel --power on" on Nelson with panel
Tested: testing on an astro with the panel
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741971
Original-Revision: 1d7d51c9d00212cb2ceb698d3795aef0993d611b
GitOrigin-RevId: a2e4ab15a68cde6f4b3aadac25be49a5ed717dee
Change-Id: Ia2d99a455e509ef5bd885c0cdfeddbe77d9c93c4