[roll] Roll fuchsia [wlan][policy] Adds tests to legacy/device.rs.
This change adds tests to a previously-untested module.

This change also fixes a small logical error in
legacy/device.rs.  Previously, if an interface could not be
added to the PhyManager, we would still try to add it to the
IfaceManager.

Original-Bug: 53900
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742745
Original-Revision: 14f3bda4fcc51c35620a575a98309ebfdce0f6a0
GitOrigin-RevId: e062e0dbd9a0428589f1a97b0540356cca5853ea
Change-Id: Ib5e090d4a3ea022b3d8d3c258988d856e2937b86