[roll] Roll fuchsia [packet] Fix options parsing bug
Previously, when parsing an options format with a multi-byte kind field,
if we encountered a partial kind field (e.g., 1 byte where 2 are
expected), we would spuriously treat this as equivalent to having 0
bytes left. This commit fixes that bug and adds a test for it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741465
Original-Revision: fd50551ba114eae1be526a567adbe73279873b6a
GitOrigin-RevId: 57dd4b9898df4c7b23625419659229623f04e96c
Change-Id: I05207d4cb2fa7a9db224ef760c83418060e3a6b7