[roll] Roll fuchsia [docs] Removing link to historic /docs page
This removes the link to historic docs/ page referenced
by a git hash. The doc_checker contraints will disallow links
to old versions of files in the //docs directory. If this needs to
be done in the future, we can revisit the constraint, or with the malice
of forethought move the markdown file out of //docs and link to it in
the new location.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742864
Original-Revision: a6ffbf5ed55536dab816205d85e3461a3a647641
GitOrigin-RevId: e12787dbf91527cccfed6e248ce452dd20a3052b
Change-Id: I2023f240e5dd1ed99f5414a9278bd6d989d13f9d