[roll] Roll infra/recipes [roll_comment] Don't comment when merging two projects
When experiences.git was merged into fuchsia.git
(http://go/no-more-experiences.git), the roll commenter added a comment
to every submitted CL in the history of experiences.git, as they all got
rolled into fuchsia.git. This caused a lot of noise on old CLs.

To prevent that from happening in the future, we should only add a
comment on commits whose Reviewed-on footer matches the `Rolled-Repo`
footer in the roll commit message, which indicates that it is the
original version of the commit, rather than a copy of a commit
originally made in another repo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/742733
Original-Revision: a7d6f7be9bd8b23c25c5bf3d97777cded63a5b4e
GitOrigin-RevId: 6f834adf12c450d9cf48bf8bba325a5b896a74b3
Change-Id: Ie802701760a501c2a5a188303d35efc259c798f9