[roll] Roll fuchsia [kernel][vm] Remove redundant user_child_count_
The user_child_count_ in VmObject is a remnant from before VmCowPages
was introduced. Before the separation of logic offered by VmCowPages, it
was possible for a hidden VmObjectPaged to exist, which meant the child
count as seen by the user and the internal child count maintained by the
VMO hierarchy could drift apart. This is no longer possible as each
VmObjectPaged has a corresponding dispatcher and so the length of the
children_list_ is the true count of children as seen by userspace.

A CQ run of fxr/742865 verified that user_child_count_ could indeed be
used interchangeably with children_list_len_ by asserting equality of
the two in various places.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742869
Original-Revision: 885d8850f612c9213d9e5f7ec4818869e429e024
GitOrigin-RevId: bdd672b864f2caea8fd2e0790425aff6dbde9e27
Change-Id: I47165e16555db6348586eecf2a2168ed1504174f