[roll] Roll fuchsia [fx][flash-remote]: check if ffx is built locally
fetch_or_build_tool is slow for most workflows, especially if the tool
is already built.

If the remote OS and the host OS are different in fetch_or_build_tool,
then use the fx-cmd-locator functions to get the host-tools dir (if it
exists) and check if the target tool exists there. If so, we can skip
building or fetching the tool and return the path to the host tool.

Otherwise, go through the fetch_or_build_tool workflow as normal.

This also adds the remote OS's operating system in the cache file
(colon delimited). Long term this will be passed to
`fetch_or_build_tool`

BUG: 79023
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732459
Original-Revision: 5def55173a2d3920959dea95f434f6b8f6c1bfed
GitOrigin-RevId: 24c92692f622f8317fad2dcace60d8fe9f898bf0
Change-Id: I547594ada181757bedcd116a3dbbf90bf622c418