[roll] Roll fuchsia [debugger] Clean up obsolete code.
Removes code protected by the LLVM_USING_OLD_PREBUILT #define. The roll
this helped support is complete so we don't need this any more.

Disables the code to set this define, but leaves it in the tree for use
next time we do a roll.

Renames DwarfDieScanner2 to DwarfDieScanner (I think we used to have an
old version of this).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742873
Original-Revision: 6741fb1b8bede88e929746b9eafa18117d1f1930
GitOrigin-RevId: 907dfe453e6589dbe4a26dde036bdfd52f4781d9
Change-Id: I171cac987e2d822084c43eeb5e2401d71ce33bc9