[roll] Roll fuchsia [audio][drivers] Add delays hanging gets
Allow for drivers to report updates to delays reported
for synchronizing presentation across the audio subsystem.
Also we add the ability to report the delay added by
hardware or software that is abstracted by the driver as a
zx.duration rather than the FIFO size, this makes the
information more generic for drivers that don't abstract
hardware with a FIFO, examples are a Bluetooth output
driver and a PDM input driver with software in the data path.

Original-Fixed: 109820

Test: fuchsia-audio-device-output-tests
Test: sa-unittest
Test: aml-g12-tdm-test
Test: aml-g12-tdm-dai-test
Test: usb-audio-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727151
Original-Revision: de4a543c9c9f0da65e29f56928702b56db173c9b
GitOrigin-RevId: dced79d5169cec86205a694ac85ef2f674f6ca14
Change-Id: I511cc620772fa760b31cf82acf79857695da048c