[roll] Roll infra/recipes [testing_requests] Upload zbi tool with test task inputs.
This will be needed for using ffx emu since ffx expects the zbi tool to
be there so it can embed the authorized ssh key into the zbi image.
Since we're uploading the zbi tool into the test task, we can let
botanist handle embedding the ssh key instead of the recipes in the
non-ffx case, which it already knows how to do if you provide the path to the zbi tool in the qemu config, so it is safe to submit this once fxrev.dev/742859 gets submitted which will add the tool to the build graph for arm64 builders.

Original-Bug: 95938
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/742052
Original-Revision: 8646673df5647a31e59a5442cbd9fe67ebf6c7b8
GitOrigin-RevId: 845b166808f67a1c2fbede11b9a20dd0ca3d2ade
Change-Id: I2aaca9b2fa76007923a77f782fadf8864e57d09a