[roll] Roll fuchsia [fdf] Move log to common location for oom test
This moves the log into a common location so the oom test can find it
for both dfv1 and dfv2.

Original-Bug: 108569
Tests: The test is still not functioning due to missing shutdown path in
dfv2 but it does pass this log check now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742855
Original-Revision: fe833a8456861a15fea374c4a02e79887af034b8
GitOrigin-RevId: 06eeab7a956c36a1f3783976f3e407f5bcdc3bf3
Change-Id: I53358a42f86dffa8d031b1a8c58ef89746eae7e2