[roll] Roll fuchsia [kernel] Print OOPS when secondaries fail to boot
Print a KERNEL_OOPS if we fail to boot any of the secondary CPUs.

The motivation is to ensure that CI/CQ detects the failure while
allowing the system to proceed so that logs and other diagnostic data
can make it out of the system.

Original-Fixed: 112387
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743503
Original-Revision: 01905d7755cf6cdbad1f671c67208d7c31876e8c
GitOrigin-RevId: 5ed8e647df954b6c76ed5787548d14bb65bff617
Change-Id: Ia6b55c965582e1bff4b4ebc788ee950442cd6cd6