[roll] Roll fuchsia [zx][kazoo] Introduce @embedded_alias
FIDL IR buggily does not (yet) preserve the name of aliases (and instead
replaces instances with a full resolution). To temporarily overcome
that, we introduce `@embedded_alias("<alias name>")` meant to annotate
`vector`s and `experimental_pointer`s with aliases as parameters: this
injects knowledge of the alias name into the associated IR element and
allows for backends to properly generate pointers to `Futex` or vectors
of `paddr`.

This change removes the last of the magic aliases.

Original-Bug: 110294, 105758
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742045
Original-Revision: 97b3ff78c853ef3a56b9f30f255baf4cee922512
GitOrigin-RevId: df2fb638ffdee531bb7821baaf67a8bd611b6600
Change-Id: Id10a7b86764eda995eb6eca1791158097e952771