[roll] Roll third_party/pigweed pw_transfer: Add lifetime retry limit
This adds a lifetime retry limit to the C++ and Python transfer
implementations. If the total number of retries over the duration
of the transfer exceeds this limit, the transfer is canceled. This
is done to avoid situtations where transfer could end up in an
infinite retry loop.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/114490
Original-Revision: 67fd10e57d821ba8efc6997f092bd27f9ae2f851
GitOrigin-RevId: b65d6ccfc98873e75b910bbf6ee71f72b4a3b918
Change-Id: I8834ce0011a56ea1b877508be721d8f5a5eda6b0