[roll] Roll fuchsia [fs_management] call shutdown in Drop
When we drop a serving filesystem instance, it's expected to attempt a
graceful shutdown of the filesystem. Fix the Drop impls in the rust
fs_management to actually call shutdown for all scenarios. Also, use the
synchronous proxy so that we know it's shut down once we finish
dropping.

Test: fx test fs_management_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742935
Original-Revision: 13a99ad2ffbfe9531f56a85bae798eef4bf4debd
GitOrigin-RevId: b54ce5fcd0d5690563344f4c839b8c453a8aacf9
Change-Id: I88738512dd54262720feb6f210aa1f374e3e806f