[roll] Roll fuchsia [netstack] Throttle dhcp logs
Applies the following heuristic to a handful of particularly spammy
log paths:
1. If we have logged, or attempted to log, this line within the past 70
   seconds, swallow the log.
2. However, if we have been swallowing the log for at least an hour,
   actually log the line (along with how many times it's been
   suppressed).
3. Additionally, we manually reset the log throttling state if we
   successfully complete a transaction.

The hope is this will silence uselessly repetitive logs that we didn't
receive a DHCP offer (especially in scenarios where we will never
receive a DHCP offer due to the lack of a DHCP server on the network,
e.g. in cases where we are running in an emulator), while preserving
sufficient debugging signal to understand the state of the DHCP client
from device snapshots.

Original-Bug: 112043
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740689
Original-Revision: d73af2c5152d1f778bb2260f33afc3286209c76d
GitOrigin-RevId: fe6484c86bbcafd9014c641c288d4d9f3fbd5f45
Change-Id: I1d30688c61e6011d5018251082dc3bfdfd8ba05d