[roll] Roll fuchsia [sdmmc] Suppress error messages for SD_SEND_IF_COND
SEND_IF_COND is not run for SDIO and is expected to fail on MMC devices,
so there are very few cases in which it is actually an error. Suppress
logging for this command to prevent error messages during normal
operation.

Test: No errors during eMMC probe on Nelson
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743071
Original-Revision: 37b87d950476dfbe3b45bc8196409d87b8f7ba3a
GitOrigin-RevId: 308435053486baa1764c469c9f2f3369aa50c44d
Change-Id: Ibbe9494c66c36ff40960f4bb86905fc6bfdcce55