[roll] Roll third_party/sdk-integration Allow for a stubbed cipd_tool
This change allows a user to stub out the cipd binary making it so we
will not use cipd to fetch an dependencies on the fly. If a user uses
this then it is their responsibility to ensure that all depenencies are
loaded locally.

To use, add the following to the WORKSPACE file

```
cipd_tool_repository(
  name = "cipd_tool",
  bin = "__empty__",
  validate_cipd_bin = False,
)
```

BUG: 112546
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/sdk-integration/+/744645
Original-Revision: 1b2d1612ac0ad098e2a1db6fcaeda1b5778804df
GitOrigin-RevId: dc5df9ee097123f5d3bba21b3cabcad209270318
Change-Id: I50c3c9fd9d3b6e755d5e66fcd7966db7fe28d1c2