[roll] Roll fuchsia [fshost][assembly] Move compilation of fshost component to assembled_system()
To allow assembled_system() to be able to properly select the features
configured in fshost, assembled_system() and fshost_cml() need to use
the same options.  This is easiest of it's just moved to within
assembled system itself, which prepares it for use with
product assembly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742562
Original-Revision: 6b0fbae6379bcb5a8e5fca747b44e90beb8045a6
GitOrigin-RevId: 9fa50b6bef4cfe7b61a41b09f8a0980f7f5d0957
Change-Id: Ibab521d0e74a30795fa2fccfac5a09cd5b86f9df