[roll] Roll fuchsia [check-licenses] gn package improvements.
The output of "gn gen" contains a list of everything that is defined in
the build. The list of things that are "defined" is vastly larger than
the list of things that are actually used / depended on in the build.

This CL changes check-licenses to now only look at the dependencies of
the //:default target, which (after talking with some folks) appears to
be the target I am most interested in for licensing. This CL adds code
to analyze that target, analyze all dependencies of that target
(recursive), and analyze any inputs that those targets may define.

The result is a significantly smaller NOTICE file, but I believe this
to be more accurate. Using `fx gn path` on a handful of now missing
targets, I confirmed that they were not a part of the build graph.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743247
Original-Revision: 3f6d61d8abcdfc60731812bffb86c0c1619e329c
GitOrigin-RevId: df13b8ecaea27436a7cb01a6ddc0b71d9ea3a23f
Change-Id: I4cdc5bc2b1db9f98371bad4695d31a2a5c65b3ec