[roll] Roll fuchsia [debugger] Add a simple bytecode machine.
This is a simplest-possible bytecode stack machine. It uses the existing
ExprValue type for the stack type. Most of the machine logic is actually
implemented in callbacks that are registered as machine operations that
take a certain number of arguments. This leaves the machine to implement
only the higher-level control flow.

Using a bytecode machine for the control flow in expressions will
eliminate the deeply nested callbacks with the code written in reverse
order, with every step being potentially asynchronous. This machine
allows synchronous operations to stay synchronous without generating
super deep stacks.

This is not hooked up to anything yet.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739606
Original-Revision: 2c632dd7eebb43c8d72797e6070966e1249469bf
GitOrigin-RevId: e54e1d8ff274b21ae7ccef1512fc868943953ed9
Change-Id: I73baaf01dec91add1732c0330c4cde87c5c3a14c