[roll] Roll fuchsia [pkg-cache] RetainedIndex supports subpackages
Before, either a package's meta.far was cached and so all its
required blobs were known, or the meta.far was not cached and
none of the required blobs (besides the meta.far itself) were
known.

With subpackages, the meta.far can be cached but not all required
blobs known if the meta.far of a subpackage is not cached.

Original-Bug: 104909
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742368
Original-Revision: 59fad43ddbc3153ce037aa99662fe71d413f0e67
GitOrigin-RevId: 06e454c5f1e4e193eba4a6fa3bb988c102895c8b
Change-Id: Ia0e65f0a702f2498d19d53cb76a3b4292fb91867