[roll] Roll fuchsia [netstack3] Add method to set RingBuffer capacity
Add the ability to set the target capacity on a RingBuffer without
losing any of the already-written or unread bytes in the buffer. If the
new capacity is an increase, the change will be applied immediately.
Otherwise the target capacity will be recorded, to be performed later
once read operations have cleared enough space in the buffer.

Original-Bug: 110625
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742644
Original-Revision: 47cbc72e901477fc161a9fb0f7f96e6e2bea062a
GitOrigin-RevId: 36317ce8f30b83b42de459ff452b633edef6d51b
Change-Id: I0277018f4f3d9a578bc54cb6606ffa617993e56e