[roll] Roll fuchsia [net-types] Make IP type a param for GenericOverIp
Hoist the target IP type for GenericOverIp up so that it is part of the
GenericOverIp trait implementation. The result is that the
implementations of GenericOverIp can have trait bounds for the IP type
that are not just `I: Ip`. This allows GenericOverIp and Ip::map_ip to
be used with types from IP extension traits.

Original-Bug: 111992
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743486
Original-Revision: feaa2f3b237444489afc1fdaae2a5044b924c93d
GitOrigin-RevId: 8735a932d1277a099c28d69030ef4d68fd91a8f1
Change-Id: Ie1762dd5372fd186b2585194ac8e972d64a1e620