[roll] Roll fuchsia [starnix] Update restricted mode debug addr
This change uses the debug address of Starnix itself for each process it
creates. This allows the debugger to get the correct module lists for
Starnix itself when using restricted mode.

Note that this is not always desirable. Work is being done to ensure
that the debugger can support module lists for both the restricted and
the normal mode code.

Original-Bug: 101486

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743343
Original-Revision: 8be640633a9b6dcdc8307b313371e4d8e1767038
GitOrigin-RevId: b42d4a5b7cc45b9f75f808efeb89d6d2a7c9c5d2
Change-Id: I625eee96f6c5fcb1382b4a07de8a95291c6108d3