[roll] Roll fuchsia [devicetree] Add MatcherScanResult
In order to have an upperbound in the number of scans we might need
to perform over a devicetree, we introduce MatcherScanResult<N>.

When a matcher may require more than one scan, it express so by
providing an overload that returns MatcherScanResult<N> where N
is the maximum number of scans needed before a Matcher reaches
completion.

Following this, for a set of n matchers, the maximum number of
scans is determined as max{ N_0, ...., N_n-1} where N_i is
MatcherScanResult<N> for matcher i or if the matcher provides
the MatcherResult overload N_i is 1.

Original-Bug: 78675
Test: fx test -v -o devicetree-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739656
Original-Revision: 49d18baa76d336e3f4ac99e6236292c4e984fe01
GitOrigin-RevId: dea1cce21e13c3e201f563527d47bc9c50431f81
Change-Id: I3d590b6dbf8b583bc0bc259073287520d66c4c2d