[roll] Roll fuchsia [devicetree] Requested Scans enforcement
Previously a matchers state was only checked if they had completed
before the maximum number of scans requested were completed for an early
exit. (E.g. a matcher requests 100 scans, but is done after the first
one).

This allows matchers to provide incorrect information when combined with
the requested scans by other matchers. (E.g. matcher 1 requests 1 scan,
and matcher 2 requests 2 scans, if matcher 1 is not done after 1 scan,
but it is done after 2, it is considered success).

This change, enforce that if a matcher requests N scans, it must be done
by the end of the given scan, otherwise return the index of the matcher
in the matcher set, so a user can debug the respective matcher.

Test: fx test -v -o devicetree-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724223
Original-Revision: c7a56a4afd68a83f042cdefa8cac91b9590197a4
GitOrigin-RevId: 3e07bebde210766cd5e6e25d2fb46026233207c5
Change-Id: I3cb2ac0f3c974028a3ee30bdcaebe370e1c07cbf