[roll] Roll fuchsia [audio][drivers][tas58xx] Allow no fault GPIO
Some product configurations may have the FAULT pin either
not connected to a GPIO input pin or used as an ADR (address)
pin and hence not available to the driver to be used for fault
detection. Allow these configurations by not failing with an
error if no "gpio-fault" GPIO is provided to the driver.

Original-Bug: 107645

Test: tas58xx-test
Test: audio-driver-ctl tone in Pinecrest with subsequent CLs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744042
Original-Revision: aeea0009f5cbf01fa03d32ab84866dcad844da54
GitOrigin-RevId: 5e6799996194a1417794909b9a6d5ed159bd17e3
Change-Id: I278dc0799b15cbf40b8d7d7831817c3443e3faa6