[roll] Roll infra/recipes [ctf_artifacts] Avoid passing sdk_id="None"
The resolved `release_version` might be None if the build was not
triggered by an integration.git commit, e.g. when triggered in
fuchsia.git CQ. Previously we would still convert it to a string
("None") and pass it into GN as the `sdk_id` arg, for which "None" is an
invalid value that would break the build, e.g.
https://ci.chromium.org/b/8799756477028251905

Now we'll skip the release version resolution process entirely if the
build is not triggered by an integration.git commit. I also added an
extra defensive check to not convert release_version to a string if it's
None.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/745242
Original-Revision: 70efff5e0a6be0953e9d009823947c8a5d178da3
GitOrigin-RevId: fa9d003389ac71a6390df8d4876c2279fd658e02
Change-Id: I3209e8bb3b00f816093d9045f8bfec75d627bc97