[roll] Roll fuchsia [make-fuchsia-vol] Prefix images with fuchsia build dir
`make-fuchsia-vol --fuchsia-build-dir ...` will look up the images
out of the build directory's `images.json` file and use them if
the paths were not specified on the command line. However, the
image paths in `images.json` are relative to the build directory.
This meant that this option could only be specified in the build
directory.

This patch fixes the problem by prefixing the `images.json` paths
with the build directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744882
Original-Revision: fd73a967991ea0450da82c21b19f3c5d78a242ff
GitOrigin-RevId: 2cd3f5c1cca720a44ee81d5becbff135c02d4093
Change-Id: I993666f8481081d0d609038ab23c093c7d41b2ea