[roll] Roll fuchsia [debugger] Use inline calls for line matches.
This works around a clang bug
https://github.com/llvm/llvm-project/issues/58483
which is that it doesn't emit line table information for the source of
an inline function call if the line consists only of that inline call.
This is fairly easy to trigger (just put an inline call on a line by
itself) and means that you can't set a breakpoint on that line (at least
in optimized code).

This works around the problem by taking all functions that match the
normal file/line lookup, and adding in the inline call information as
potential line matches.

This produces slightly surprising result as the breakpoint set is shown
to match the first line of the inline call (since this is where the
actual address matched to), but everything works as expected.

Original-Bug: 112203

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743902
Original-Revision: 66dfe93db20d295b15ffcea5fd159d933a0bb9cf
GitOrigin-RevId: 692caa58b67bca06decaec00c37d62eea54d10ea
Change-Id: Ic3e8da927e7cb16ab733f5c7f1d13570991fa9df