[roll] Roll fuchsia [emu] Multiplex monitor and serial to the console
As described in the attached bug, when running the emulator in console
mode, some command sequences were intercepted by the launcher instead of
being passed through to the guest. For example, pressing Ctrl-c would
cause the emulator process to terminate, when the expectation is that
the command sequence should be passed to the guest, and the guest would
interpret it and cause the guest shell to terminate instead.

This CL adjusts the configuration of the emulated serial port, such that
it is multiplexed with the QEMU monitor on the same channel. The effects
of this are two-fold: first, control sequences are now passed through to
the guest as expected; second, the user can utilize both the QEMU
monitor and the Fuchsia shell from the same host terminal.

The multiplexed channel also supports additional QEMU command sequences;
of particular importance: Ctrl-a c will swap between the two consoles,
and Ctrl-a x will terminate the emulation. This is consistent with the
QEMU default behavior, as well as the behavior when using the legacy `fx
qemu` and `run-zircon` launcher scripts.

Original-Bug: 111907
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744384
Original-Revision: e5d2147dafea5311f595cd687607008123e5a512
GitOrigin-RevId: 079f4384f64291224a0ab28e3e0c388bf172c146
Change-Id: I06a8da39424f9a50dad44e229c4654a50828f837