[roll] Roll fuchsia [kernel][arm64][mmu] After modifying PT entries, issue a DSB instead of DMB
DSB ISH (and the slightly weaker ISHST) is necessary to ensure the
previous writes to any page table entries are flushed out of the core
and visible by other processors and page table walkers.

Previously was using DMB which is actually insufficient since it only
guarantees the order the entries are flushed into the memory subsystem,
not that they actually are.

In most cases it's sufficient to use a DSB ISHST since it ensures only
previous writes were barriered, and reads by the cpu can go around it.
This is okay because the code wont access any of the new mappings
directly until at least exiting the mmu code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669549
Original-Revision: 01bf4889b62c649af6f8397fbe3619e52e23b0ae
GitOrigin-RevId: 86c46a6cea22e6f97aaebc0f40b11a5e24130d1b
Change-Id: I87cd033b4adcf1e4757ee62851356e6c04d47fed