[roll] Roll fuchsia [driver2] Add prepare_stop hook
DriverRecordV1 was renamed to DriverRecord and reorganized to make it
clear which members belong which version in the same style
async_dispatcher_t is defined in libasync.

New macros were added for supplying the prepare_stop hook and a default
implementation was added to driver::Base.

Test: fx test dfv2_v1_lifecycle
Original-Bug: 97457
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724404
Original-Revision: 6db867558369a60d0e9173c6ce51724bb2214c16
GitOrigin-RevId: d8325f060b74915ad99a191ad5ba33ff98ad635d
Change-Id: I41cac95b7505279a5483c2a15ad317e885eb9d04