[roll] Roll fuchsia [debugger] Emit bytecode for the bytecode machine.
Emits bytecode from the ExprNodes to run in the new bytecode stack
machine.

No attempt was made to share logic between the AST execution model and
the bytecode model. This resulted in significant code duplication
between the two models. The reason is that if successful, the AST mode
code will be deleted, and if unsuccessful, the bytecode mode code will
be deleted. Since this transition should not take very long, duplicating
the code allows one or the other mode to be deleted with minimal
disruption.

This code is not yet hooked up except for tests. The ExprNode tests now
run each variant of the test in the old "Eval" method that runs on the
AST, as well as the bytecode version.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739607
Original-Revision: 709e69fc98e4280246f79030e58e6591560280b5
GitOrigin-RevId: aff78fd6860ba55a2df25cc7d4e0401f2fa03978
Change-Id: Id7de0d15e6c71b931ff12731c553f4449b18fe56