[roll] Roll fuchsia [kernel][x86] AMD Zen 2: Mitigate RETbleed (STIBP, SuppressBPOnNonBr)
This is a _partial_ mitigation for RETbleed / Branch Type Confusion
on AMD Zen 2 processors. Family 15h and Zen 1 processors are not
mitigated in this commit.

AMD Family 15h, Zen, and Zen 2 processors are affected by RETbleed [1]
and a generalization, Branch Type Confusion [2]. Processors may
mispredict the _type_ of an instruction (No Branch, Direct, Indirect,
RET) and/or the target of the instruction; there is a window between
the misprediction and a redirect to the correct type and target. Type
mispredictions may be caused maliciously and the window between the
misprediction and redirect may be exploited to bring data into caches
and to leak or infer the contents of that data.

Only certain types of type mispredictions (actual type vs predicted
type) may have a window large enough to exploit. See [2] Table-2.

0) Enable STIBP - Single-Thread Indirect Branch Predictor
   This by itself doesn't prevent Branch Type Confusion, but is a
   prerequisite for any BTC-RET mitigations (jmp2ret or IBPB on
   kernel entry).

1) Set MSR DE_CFG2[SuppressBPOnNonBr] to mitigate BTC-NOBR cases

Fuchsia already had mitigations for certain BTC cases; these were added
for BTC and for prior speculative execution issues; see [2] table-3:
- fxrev.dev/693722 and fxrev.dev/661362 added INT3's after RET/direct
  jumps ('SLS protection' in Table 3)
- Fuchsia already uses Retpolines on AMD Zen / Zen 2 processors
  ('IBRS/Retpoline' in Table 3)
- Fuchsia already enables SMEP & stuffs the Return Address Stack
  on context switches ('RAP protection' in Table 3')

Remaining cases / mitigations required:
- BTC-RET (Mispredicting the type or target of RET) - jmp2ret or
  IBPB on kernel entry;
  In addition, defense-in-depth mechanisms may make it difficult
  to exploit BTC-RET - fxrev.dev/710262 tracks zeroing used
  registers before RET, limiting state an attacker can control.
  See [2] section 7.1
- BTC-DIR (Mispredicting the type or target of direct branches)

[1]: https://comsec.ethz.ch/research/microarch/retbleed/

[2]: Technical Guidance for Mitigating Branch Type Confusion,
     Revision 1.0, 2022-07-12
https://www.amd.com/system/files/documents/technical-guidance-for-mitigating-branch-type-confusion_v7_20220712.pdf

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739984
Original-Revision: 3d0d8e14c2aaab3000576497e2854663a22bfec7
GitOrigin-RevId: efaafce98138e57b3440bc382b90f7df1aa04d3c
Change-Id: I5d44a0c71e7d07b849b6df9d3dbd0bc573d29ac8