[roll] Roll fuchsia [superproject] Roll fuchsia [virtualization] Parameterize net device tests
This runs the tests against the C++ and Rust components. This skips the
RX test for now, as the device doesn't yet support that.

This also does a cleaner shutdown (which the C++ device doesn't seem to
do). This stops some asserts firing when the component terminates, which
only happens for the Rust device because we support stopping the
component on error.

Original-Bug: 95485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743964
Original-Revision: 1bfa5738ae63205d0de3322369c59484a74c22c0
GitOrigin-RevId: 72531e5ed16117529b58dc8e8974c8fc401a5758
Roller-URL: https://ci.chromium.org/b/8799438882802992673
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746042
Original-Revision: e2adc3b47f976948922e637f1b02697b45666ea8
Change-Id: If276cb92a108b6527cb7ae5cf3ee47c3e4fcb4e8