[roll] Roll fuchsia [netstack3] Avoid panic in TCP
When binding a TCP socket, we used to first remove the unbound socket
and then add a bound entry in the socketmap. However it is hard to hold
in a later change, we added a possible early return after removing the
unbound entry and adding the bound entry. This caused panics in syscall
tests. Now it is rewritten so we only remove after we know the new entry
is successfully added. Fixing this issue exposed another potential panic
when we didn't cancel a timer.

Test: added a unit test; manually ran syscall tests locally that are not
run in CQ yet, verified that none of them will panic the netstack3 now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745358
Original-Revision: a86d904982ed6c8ce06ec53d5d75a95ad7695919
GitOrigin-RevId: 3e4312f5419d68cb024254afd80246d6ed42e396
Change-Id: I6259fc625a4a695f7db3efa20667585f207a646d