[roll] Roll fuchsia [nxpfmac] Fix deadlock in EAPOL response handling
There was an issue with EAPOL response handling where it synchronously
acquired a lock that could be held by another call from SME. At the same
time the EAPOL response code would block the driver from servicing the
SME request. This resulted in a deadlock.

The code already performed the EAPOL response call to SME in a separate
thread. This change makes sure that the validity check of the SME
interface along with holding the lock required for the check also
happens in the separate thread.

Test: WlanSecurityComplianceABGTest ACTS test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745985
Original-Revision: 27f9696a84508539b88ef769d7b04b04bc379fba
GitOrigin-RevId: 9d972e37f1315799cfeeb59a7643d85e7af0c00b
Change-Id: I1454c6eecc001316d8709de42112d7b3d33fda64