[roll] Roll fuchsia [nxpfmac] Properly support WEP
The driver had an initial attempt at supporting WEP but this was
insufficient to make it work.

This change adds the additional call needed to enable the WEP key when
configured as well as removing incorrect IE configuration and handling
the detection of shared or open WEP keys.

BUG: 108742
Test: WlanSecurityComplianceABGTest ACTS test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745986
Original-Revision: 1049c03b54006ffd3889e5f52905f2467258e72b
GitOrigin-RevId: 223d174cc37a2e55133e5e5bf6a5b36dbf237070
Change-Id: Ib39d6cc86de18e3d32ee3991605b6daaa266ae56