[roll] Roll fuchsia [credential_manager] Improve Error Types
CredentialError is a FIDL type which is intended only to be returned
to callers of the `credential_manager` service. This type of error
is specifically intended to return actionable information to the
caller but is not suitable as an internal error type for debugging.

This change utilizes thiserror::Error to build out a new rich error
type called CredentialManagerError. This rich type is an enum that
supports transparent passing of error types from the various
subsystems such as the LookupTable and the HashTree. This approach
removes almost all the map_err(..) code from the code base which was
causing a lot of information loss in debugging instead allowing for
the rich errors to be passed through.

This should be a significant quality of life improvement when
debugging the credential manager and also significantly cleans up
the code base.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741246
Original-Revision: d0d3ea13b0abac65bc606d8d6917c6ed52062dcb
GitOrigin-RevId: b7be276240ac8c45ffe79316ba6c58f65095d6e9
Change-Id: I4fe4c1eb15675fa1ea3116b7245d291b3aafc563