[roll] Roll fuchsia [dhcpv6] Immediately use Advertisement w/ max pref
According to the DHCPv6 RFC, when a client receives an Advertise message
with the maximum preference value of 255, the client is expected to
immediately transition to the Requesting state to attempt to have leases
assigned by the server that sent the Advertise message. The RFC does not
give room for the client to choose not to use the server if the server
is not able to give the client everything it wants (addresses, prefixes,
DNS servers, etc.). Instead the RFC requires clients to unconditionally
use the server.

Before this change, the client would only accept an advertisement with
the maximum preference value if the advertisement gave the client
everything it wanted. This change updates the client to be less strict
and instead, accept the advertisement as long it has at least one IA
(even if the IA doesn't match the client's hint or the client requested
more IAs than what the server advertised) to align the client with the
RFC specification.

As per RFC 8415 section 18.2.1,
```
   A client MUST collect valid Advertise messages for the first
   RT seconds, unless it receives a valid Advertise message with a
   preference value of 255. The preference value is carried in the
   Preference option (see Section 21.8). Any valid Advertise that does
   not include a Preference option is considered to have a preference
   value of 0. If the client receives a valid Advertise message that
   includes a Preference option with a preference value of 255, the
   client immediately begins a client-initiated message exchange (as
   described in Section 18.2.2) by sending a Request message to the
   server from which the Advertise message was received. If the client
   receives a valid Advertise message that does not include a Preference
   option with a preference value of 255, the client continues to wait
   until the first RT elapses. If the first RT elapses and the client
   has received a valid Advertise message, the client SHOULD continue
   with a client-initiated message exchange by sending a Request
   message.
 ```

Original-Bug: 80595, 112643
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745446
Original-Revision: 11164983a662d40d553b0076cba8070b53d1b7c0
GitOrigin-RevId: 2517f770610284d2822acd81b44f63b535fa2e63
Change-Id: Icd5c47ed557b2ed8ae775db4bd03f88d238b43df