[roll] Roll fuchsia [line_input] Keep raw input mode when prompt is hidden.
Previously the line_input would revert the terminal mode back to the
original one when it is hidden to allow output to be printed like
normal. But this also put back normal input mode so control-C
would exit the program (rather than call the associated callback) and
typing would be echoed to the terminal, even without being processed.

This change creates a new "hidden" terminal mode. This mode keeps the
raw input flags so input is processed consistently. It now also
explicitly sets the output flags we expect for "normal" output so it
will be correct even if the terminal was set to a weird mode when the
process starts.

Original-Bug: 112841
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745347
Original-Revision: 1daefa6f5bf0c328d54f5c08e2cb251a22f41705
GitOrigin-RevId: b55a997cb7b2be4326583ace45db10cdaaafb0cb
Change-Id: I303a09c3dc79f48109ce5a1208ad1d704a9120c1