[roll] Roll fuchsia [debugger] Add thread controller completion CBs.
Thread controllers have a callback on init so that the caller can know
if the controller initialized itself properly (useful when there are
requirements about breakpoints needing to be set).

But there was no way to know when a controller thinks it's "done." This
adds an optional constructor callback so the creater of the controller
can observe this event.

This callback is currently unused and there should be no behavior
change. It will be used in a follow-up CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744786
Original-Revision: d7178101a936998968ec39ade4a9ecadeca6416f
GitOrigin-RevId: 44c33476c5aab85956248a584d41d69ef4848193
Change-Id: I05d18c6d72852d8e0d9a57d29b28bca79c4392c1