[roll] Roll fuchsia [netstack] Disable NS3 ifwatcher test variant
The NS3 variant of `test_watcher` was accidentally running with NS2.

The test does not trivially pass against NS3 due to differences
between the two netstacks. A bug has been filed to split this test up
so it is easier to enable against NS3 and account for differences
between the Netstacks and a TODO added to address this.

Original-Bug: 112627
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745088
Original-Revision: e0e2d1b041eb07a16e8379c4872016ce60f44ce9
GitOrigin-RevId: 6e8ce128ec05e36c6a0aabccf86dd2514f0af08b
Change-Id: Ib83f42be088fd4d7c27a6bcaf9fdf9ed235e31f0