[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] add performance options for string to float
This allows the client to set compile flags to disable the passes that
the string to float function uses. A client may be willing to trade off
performance for a reduction in code size, and this allows for that
fine-tuning.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D136495

GitOrigin-RevId: 9a60e63dfe083396c02fb6c0c8e3eff4e45ae9dc
Original-Revision: 31449b5cf19c9f3152026abbdb1836345e1a0090
Roller-URL: https://ci.chromium.org/b/8799386778149413361
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744571
Original-Revision: 307e520e070a97f22faf382ac233a1661f6dfa51
Change-Id: If141964f6785ea5f39acbdf89a5970b63af535c5