[roll] Roll fuchsia [zx][kazoo] Remove @arg_reorder
This change replaces @arg_reorder magic with a clearer decomposition of
would-be buffers into separate pointer and length parameters. In the
cases where we want to treat the disjointed pointer as an out parameter
(amidst in parameters), we introduce the @out annotation to signal that.

Original-Bug: 110294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743087
Original-Revision: b2ad72ecb82b4985288520392d4924fb7eff1a26
GitOrigin-RevId: a2edf6bd95443e40135278403788b53a0ef0eb59
Change-Id: I290e88b186af7ebccb8b01434d857242906a4573