[roll] Roll fuchsia [test_manager] Don't crash if logger connection fails
Return an error from the hermetic resolver instead of crashing the
entire process when connecting to LogSink fails. This connection error
is expected in cases where the hermetic resolver is running before
Archivist but the test realm is being shut down. Because hermetic
resolver is a mock and runs as a Rust task on an executor, it could run
for a little while even after stop is called. As a result, it may
attempt to connect to LogSink in this situation, and the connection
could fail.

Original-Bug: 112984
Multiply: run_test_suite_integration_tests, test_manager_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746322
Original-Revision: d6d02c156512645fea85efd1facad41c674cdda5
GitOrigin-RevId: 44faefea04f2949416065bb57825cf82cc852975
Change-Id: Ibcfb3404ce0d3ba569a41ebfc17d648550151fab