[roll] Roll fuchsia [ffx][test] Update help output
Update the help output for ffx test run in preparation for running it
through a design review. Reordering the options within the struct is
relevant as the order in the source code determines the order in which
the options are printed in the help message.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743082
Original-Revision: 065b7615ab17e0d5fb462bcec0a4b7980ccb3ec5
GitOrigin-RevId: 4c583103f36a6bb57fff7c4d4f2cf51c8c0dd4cf
Change-Id: Ifb5898c156740ea74662369b732f1bfcf6ffbed4