[roll] Roll fuchsia [display] Remove unnecesary device controller
This controller handle was being passed to the DDK function
add_device_controller, which the Driver Framework is trying to remove.

It looks like the only use of this handle was to keep the device alive.
If this handle is never passed in the first place, the device will be
kept alive by the controller handle.

Most of the comments in the removed code suggest that this device handle
is entirely unused.

Original-Fixed: 33675

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744863
Original-Revision: 3803cb8bab8fa1f92ae3713b7bfd78ad6dd0dcb2
GitOrigin-RevId: 85f519be10df8a3b231f4d233d3331c995dd6b29
Change-Id: Ic44940d76f6a7040998d1c71c44edd3f7e4c59c1