[roll] Roll infra/recipes [yasm] Recipe for building Yasm
We use Yasm as a replacement for NASM. Currently, it's always built from
source as needed, but that's wasteful since we do a lot of duplicate
work. We would like to instead setup a builder to build Yasm and push it
to CIPD as a package that can be consumed from builders that need it.

Original-Bug: 108206
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/745424
Original-Revision: add786eaf4c5473c1b4b03693e9b7dffb94d1b9f
GitOrigin-RevId: 41114863c56ba8c66c981083a290c8caab08355e
Change-Id: I023cf9010c902b1a99967b8fff6fb99fc7796e36