[roll] Roll fuchsia [component][fuzzing] Update GN templates for libfuzzer
This is mostly a reland of commit
8a9594292d92e118b468ddb5b18d65e5bc53009a

It differs from the original in that `corpus` has been moved to be a
dep of a group that includes the fuzzer executable, rather than a dep
of the fuzzer executable itself.

fxb/113012 discovered that when the corpus relied on being built by
ninja, e.g. as the output of a compiled action, it could result in
ninja subsequently looking at the executable and deciding its inputs
had not changed and it did not need to be rebuilt. This in turn could
lead ninja to drop actions that solely depended on that executable
target, including the packaging step that included the corpus.

This manifested in a ninja no-op check failure when it noticed
package-related files that were older than the generated corpus that
the package contained.

Original change's description:
> [component][fuzzing] Update GN templates for libfuzzer
>
> This CL adds new templates for building library fuzzers in the v2
fuzzing
> framework. These templates try to present the same "look and feel" as
> those for components.v2, e.g.
>   `test()`                   => `fuchsia_library_fuzzer()`
>   `host_test()`              => `host_library_fuzzer()`
>   `fuchsia_test_component()` => `fuchsia_fuzzer_component()`
>   `fuchsia_test_package()`   => `fuchsia_fuzzer_package()`
>
> Original-Bug: 45347
> Original-Reviewed-on:
https://fuchsia-review.googlesource.com/c/fuchsia/+/706926

Original-Bug: 45347, 113012
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744580
Original-Revision: 6f05539307428e63ba5d4e877fbb39c8f1ba117e
GitOrigin-RevId: 7c40414c58ffda6d57addd49d95cb7fd96bb84bf
Change-Id: Ibcd888a3e98ce90dbe2f210325f676f847b06c34