[roll] Roll fuchsia [kernel][hypervisor] Avoid permission faults
The hpfar_el2 register, which is relied upon to give the IPA to resolve
a data or instruction abort, is not valid for permission faults.
Permission faults therefore cannot be handled, and must be avoided.

This change modifies the protect path for stage 2 aspaces to perform an
unmap instead, and adds checks to the instruction and data abort
handlers to ensure the hpfar_el2 is valid before attempting to use it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744825
Original-Revision: 74999e3abe8c88c0f040929ffa8d046e3b421f72
GitOrigin-RevId: 76124bccf55152ee1472888f42809784ff813382
Change-Id: Ic3ee572440290c726404bf48df96db6d9a8c85df