[roll] Roll fuchsia [ktrace] Migrate VCPU Events to FXT
Because this means we no longer do a post processing step, we don't know
the exit type of the event when we emit the vcpu begin event. To work
around this, the exit type has been moved to become an argument and the
even name is now just "vcpu".

Original-Bug: 98176
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743067
Original-Revision: b3b09389d9ebeed10383bdbaef5e7816eea1c340
GitOrigin-RevId: 4ebf680e198e704ba7570021b37838b8e9357579
Change-Id: Ief7636dcb84157f07a6439e028b1b2a8076ce509