[roll] Roll fuchsia [ui][flatland] Add display controller rotation.
ImageRect:
---------
This is achieved by creating a new struct in Flatland called ImageRect.
The major differences from Rectangle2D:
- It holds unnormalized unrotated UVs
- It holds the Orientation value.

Calculating Orientation:
-----------------------
Instead of decomposing the matrix, we already need to calculate the
reordered_verts as part of calculating the new origin and new extent.
The reordered_verts have the same sequential ordering of verts, but
are offset depending on the rotation applied. We use the offset and
calculate the offset.

Calculating the Clipped Texture Coordinates:
--------------------------------------------
I'm sure the math here could be simplified :). Let me know if you find
a more elegant readable solution. The crux of the issue is that the
texture coordinates are unrotated, whilst the actual clip region is
rotated. I tried to provide details in the COMMENTS though the
calculation looks complicated and may be hard to read. Let me know if
this doesn't make sense -- I can provide more detail offline and try and
spruce up the comment further if so.

Testing:
-------
- Tested on workstation_eng.qemu-x64 to ensure nothing is broken with
  and it remains useable.
- Unit tests for display compositor added
- Unit tests for global_matrix_data fixed: Rectangle2DTest ->
  ImageRectTest
- Note that renderer_unittest already seems to have some end-to-end
  tests for both ScreenCapture AND vk_renderera (see
  RotationRenderTest). I relied on this test to ensure there were no
  issues with either of those two components. Adding more tests to
  ScreenCapture seemed out-of-scope for this CL.

Original-Fixed: 108821

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739579
Original-Revision: d9749c877bdafea2aad87361cff8792d13cd22a5
GitOrigin-RevId: 87fb8b6f4ea185de9ff91221adea6638cc36ac38
Change-Id: I7d3b3fb97136e1888fc20ae9a78007ed9ac451b0