[roll] Roll fuchsia [vulkan] Wait for old command buffers before importing semaphores.
New versions of the Validation layers give this warning in
SwapchainFidlTestSuite/SwapchainFidlTest.PresentAndAcquireNoSemaphore/UnprotectedNoCopyValidationAfter:

Cannot call vkImportSemaphoreZirconHandleFUCHSIA on VkSemaphore
0xa000000000a[] that is currently in use by a command buffer.

This is because we wait in acquire using the zircon semaphore, so they
don't realize that the command buffer must necessarily be completed.
We can signal a fence in the present submit and wait on that before
doing vkImportSemaphoreZirconHandleFUCHSIA to ensure that the validation
layers are aware of the behavior.

Original-Bug: 112606
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745464
Original-Revision: e9f9d7c33f35f0b7c35075c44e4b7ab6ea7799e8
GitOrigin-RevId: 4cf2be9343095f7104082ebc2c88dfe4d4edf760
Change-Id: I5221dbefb3ee4a8cead15d9a19dc971977c98211