[roll] Roll fuchsia [dev-ec] Add traces for lid close events.
Recently there have been a number of reports of devices not shutting
down when the lid is closed. The EC sees lid closed events however the
lid driver does not. This bug is difficult to reproduce so we're adding
some traces so that next time it happens the logs will be more helpful.

Original-Bug: 111460

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745047
Original-Revision: 39d85bd61d360db3421947a335d75dc41d3eddce
GitOrigin-RevId: b78d8d193a9416f7b0085e29a1835e9b6d87a534
Change-Id: Ie7bfd1633709dbadf3f8b4dd4b31108ecf92dd8a