[roll] Roll fuchsia [vaapi] Set min_bytes_per_row for chromium
Currently the ImageFormat bytes_per_row field is not used which means
that the codec can only change the stride when new buffers are allocated
and the constraints are set. Instead of using bytes_per_row we must set
the min_bytes_per_row as a part of the output constraints and then stick
to that value so that we and chromium agree on the buffer stride.

Original-Bug: 112011
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746703
Original-Revision: 3b2a478901bd88e7399f9710a3c657501a101fe3
GitOrigin-RevId: a5ecb86abe8421c44b38606ef3e769b9f5d3811a
Change-Id: If42bde742b681b0b8e6288e78b2b2a08df0e31c0