[roll] Roll fuchsia [starnix] BinderProcess must not outlive the given process
BinderProcess was owned by the fd that created it. Because the fd can be
duplicated and passed to another process, it allowed BinderProcess to
outlive the process itself. Also, it allowed another process to access
the fd.

To prevent this, this CL changes the fd to only contains the pid of the
process. On each access the process is retrieved from the driver, and it
also checks that the pid is the one of the calling process.

Original-Bug: 101486

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745143
Original-Revision: 2b3f51d2dac670afd50516bde20edb5ba38c87fa
GitOrigin-RevId: 4ebbbe9b50fc3d7e960e2d1d33f9e0c1cb6dbbbc
Change-Id: I26a7a17626be1fab0a4bbeff5e3df589973d71e6