[roll] Roll fuchsia [debugger] Make Symbols know their DIE offsets.
The original design for the symbol system object model was that the
LazySymbols have opaque identifiers used by the SymbolFactory. For the
DwarfSymbolFactory production implementation, it uses DIE offsets. And
once you go to a Symbol object, these DIE offsets are lost. This makes
it impossible to go from a Symbol back to a LazySymbol, or to compare
symbol identity using DIE offsets.

It keeps coming up that we need to either create LazySymbols or compare
symbol identity. The opaqueness of the factory data in the LazySymbol
has also not proved particularly helpful. We almost never use the mocked
SymbolFactory and even when we do, creating synthetic DIE offsets is
fine.

This change changes the name of the opaque "factory data" in the lazy
symbol to be an official "DIE offset" and makes it accessible. It also
adds this information to the Symbol object, allowing the Symbol's unique
DIE offset (if not synthetic, see comments in symbol.h) to be compared
and allowing LazySymbols to be created that reference the "real" object.

Changes the MockSymbolFactory to a wrapper object that explicitly cleans
up the symbol references to prevent leaks from reference cycles.

Adds the DIE offset to the debug printing of symbols. This will help
debugability of the debugger since it is now possible to precisely
correlate the output of sym-debug commands with the corresponding items
in the DWARF file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744575
Original-Revision: 01ad03f6046ff3c43fb9d45dc1cf48bca524bb06
GitOrigin-RevId: dbb72d405a8c7e72189f93ef5c085585216ba2e4
Change-Id: I6ae9b0093a95e3883516ae4973946e4f2de730dc