[roll] Roll fuchsia [debugger] Don't return LLVM types from LineTable
Previously the DwarfUnit and LineTable classes returned an
llvm::DWARFDie object. This is a layering violation as we prefer to keep
the LLVM types contained in the low-level symbol factory
implementations, and there's not much one can do with this structure.

Now that the symbols expose their DIE offset and can be created from
this offset in the symbol factory, we can replace these LLVM types with
a simple symbol offset. This also lets us make some functions in the
DwarfSymbolFactory private.

This lets us clean up some references to DwarfSymbolFactory and replace
with the abstract SymbolFactory base class.

There should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744576
Original-Revision: 65628280dc77d6833e83188e5edcfe7a9660cac8
GitOrigin-RevId: 1119e82e0c95c5ed05c546d73b9a9df6405d0d98
Change-Id: I9797e78ff5e11327105bfd41bc1fa448d8561236