[roll] Roll fuchsia Reland "[sysmem][fuzzing] Update fidl-c allowlist"
This is a mostly a reland of commit 543edaf5d2823d7eb08b2d892cecbd9099805255.

The only change is to update the names in the deprecated C bindings list
to match the chaneg in the `fuchsia_library_fuzzer` in fxr/744580.

Original change's description:
> [sysmem][fuzzing] Update fidl-c allowlist
>
> This CL switches the sysmem fuzzers to use the v2 GN templates and
> updates //build/c:fidl-deprecated-c-bindings to match.
>
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706923

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747542
Original-Revision: 066d7d6fcf059816f9d97455769670141385ee5d
GitOrigin-RevId: cd19051628cb373b08e781a4bce903f05d816403
Change-Id: I7c012918c6c21da69176b3139e643fd714b671c2