[roll] Roll fuchsia [elfldltl] Add filesz() to ZeroFillSegment
This makes it much easier to get the filesz from a segment variant
by using a visitor like `[](auto &&s) {/*something with s.filesz()*/}`

The alternative would be to have an overloaded operator() for each
segment type, which isn't that ergonomic for something that should
be simple. The cost of this method is also basically 0.

Original-Bug: 68283
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747036
Original-Revision: 36d18f549f5dbb8f7422bb6ea07d9253edbb1f31
GitOrigin-RevId: cb82e9767dfa8f4de5ca0dc908f577746f3086a0
Change-Id: Ic35f9ec72b63c558252d6660a966c979ecd82385