[roll] Roll fuchsia [netstack3] Make TCP `handle_timer` more strict
Currently, `handler_timer` is lax about if the connection ID given is
valid or not and only do work if the ID is valid. This has papered over
some other bugs. This CL strengthens the precondition of the function
so that we can catch those earlier.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745523
Original-Revision: a78b8b3239d4208768c0fd98045f63db0345d68a
GitOrigin-RevId: 9bce4ac59e4499afe817dbc03952617ca578c4d3
Change-Id: I9ddea93666f73c48fd740e56f6c590ebd11c1640