[roll] Roll fuchsia [vaapi] Reset decoder on stream stop
If the stream is stopped then the decoder can reset its context. For
tiled decoder which DPB are sysmem buffers, this matters because
CodecImpl can call CoreCodecEnsureBuffersNotConfigured() after the
stream is stopped and if the decoder's context is not reset it will
still have reference frames that are backed by sysmem output buffers.
This can cause ASSERTs to fail since it is expected that all buffers can
be deconfigured after a stream is stopped.

Fix: 110813
Original-Bug: 111139
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739584
Original-Revision: cd657a3c77357596db407788d7e6726fefd3ef10
GitOrigin-RevId: 0e53da4ea7c198ab881b6da83ad7407e7df95940
Change-Id: I401c95c50270e2d291caa49a8aa8c6e485b2a8a8