[roll] Roll fuchsia [debugger] Add option to make input synchronous.
Adds a global setting "ui-timeout-ms" which, if nonzero, will block the
input for the specified timeout waiting for a command to complete. If
the command finishes within the timeout, the UI will work synchronously.
If the timeout expires before the command does, a message will be
printed and the prompt will be shown again.

The current implementation sets the timeout to 0 which disables the
optional new synchronous behavior and keeps the behavior the same. We
can enable this manually to test until we're confident that this is the
behavior we want.

Update some commands to use CommandContext that were previously missed.

Change the commands that take thread controllers to bind their
CommandContext objects to to the lifetime of the ThreadController. This
suspends output for as long as the corresponding step command takes.

In testing this seems to work well when the timeout is set to nonzero.
The only problem I noticed is that input while the UI is disabled is
discarded. It would be better to buffer this.

Original-Bug: 109133

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744787
Original-Revision: 981d73cef2446690fc846e19be15766efe49648c
GitOrigin-RevId: 8a4480d1f9e7280c1a5084d599e6d82a96ba2f08
Change-Id: I44adb5c44ba88b746d0cbec82247a20cb4debc74