[roll] Roll fuchsia [build] Stop publishing <hash>.root.json
We will be migrating to a rust-tuf based system for creating TUF
metadata, and so we'll no longer produce the TUF-0.9-era
<hash>.root.json file. We never used this file on a production system,
so it should be safe to remove this file.

This also changes the system to copy the `root.json` from the build,
rather than inferring it from one of the root metadata files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747951
Original-Revision: 578d5f62cd15e9c5b8bb5200d343680815b6a626
GitOrigin-RevId: b6c77c3f13f8112252ec17f52ef339b31473c501
Change-Id: I4a69e1c803735881bef69e5cabc93bdda1daa718