[roll] Roll fuchsia [brcmfmac] Refactor roam engine setup code.
The Broadcom firmware has an internal roam engine, which is a firmware
offload that can be enabled and configured to make roaming decisions.
brcmfmac currently has this offload disabled. Ongoing work to support
the BSS Transition Management (BTM) offload uncovered that the internal
roam engine must be enabled in order for the BTM offload to work. This
change adds a Broadcom feature ROAM_ENGINE (disabled in the driver) that
will allow us to easily enable and configure the roam engine for ad hoc
builds and for unit testing. This change also endeavors to make the code
that controls whether the roam engine is enabled/disabled less
confusing: the iovar that controls it is unfortunately named "roam_off"
which has been cause for confusion (because a true value for the iovar
means that the roam engine is disabled).

This is a refactor of a disabled feature, which was not enabled in the
driver previously and is still not enabled in the driver after this
change. A few unit tests were added to verify that relevant iovars are
being set correctly by the refactored code. This change passes ACTS
suites VapeInteropTest and WlanTargetSecurityTest.

Original-Bug: 111760
Test: fx test brcmfmac_sim_test brcmfmac_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740509
Original-Revision: d7c34ae96f708f4e1189ce81d1da2a365c6f278b
GitOrigin-RevId: 6b0a4a7a3e20f52ac2420130c991b534f926f98e
Change-Id: I3c1ba95ac16ff2278d295804cb0e40909fa757d8