[roll] Roll fuchsia [nxpfmac] Handle deferred handling event
Mlan can send a defer handling event which means that we need to trigger
the main process when it arrives. This happens for things that are
initiated by our driver where mlan needs additional main processing to
happen for the operation to complete. Talking to firmware for example.

Handle this event and call on the bus to trigger main processing when it
happens.

Test: Included unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747612
Original-Revision: 35463282b84956c5ab10db8869f88d0a7cea2a5d
GitOrigin-RevId: 281629dca75857d4382a83af7829ceb62be51cb6
Change-Id: I19450d93334d6d875adad5ca7d25dfb213505c52