[roll] Roll fuchsia [sdk][gpio] Add FIDL metadata for GPIO init driver
Add metadata to be passed from board drivers to the GPIO init driver.
This metadata consists of a vector of GPIO fragment names and the
operations to performed on them during bind. Devices that require GPIO
configuration will bind to the init driver to ensure that everything is
set up ahead of time.

Original-Bug: 113037
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744736
Original-Revision: 2fdfed59691e5099215f9de975c2eb006c38081f
GitOrigin-RevId: d68f770e190dac677177ab9a80ae9131751a4537
Change-Id: I9e52c452325771184b0ad09ac5de7208e036a03a