[roll] Roll fuchsia [vpython][yapf] Remove -vpython-interpreter from vpython invocation.
In a decision made today regarding the future of vpython in-tree, we've
decided to, in the short term, allow vpython's python version to diverge
from Fuchsia's vendored python.

The reason this is acceptable with respect to RFC-129 is that, in the
latest vpython, it uses a bundled python executable rather than
traversing $PATH for one, so it doesn't cause any nondeterminism in the
python version.

In a change to be merged shortly after this one, we'll roll the latest
version of vpython in-tree.

Original-Bug: 107761
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748102
Original-Revision: 1d7a42e62ee1679445bcc207081266fa572d08f9
GitOrigin-RevId: bb0049866fa674d42ba2650f7f57ae75893f6ca7
Change-Id: I8578d3aa9f4050a41911ea9316e3227574260f1e