[roll] Roll fuchsia [nxpfmac] Fix IoctlAdapter timeout edge case
There was an edge case where a timeout could schedule a cancelation at
the same time as a completion was in progress. This lead to a double
call of IoctlComplete which in turn attempted to use the callback twice.
The code already protects from this by taking the callback pointer out
of the request, it can't be called twice. However the second attempt
would crash.

Fix this issue by checking if the callback is still available, if not
the ioctl complete turns into a noop and the actual completion wins.

Test: Included unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747613
Original-Revision: 96f6290ffda94ea88ee0baadac4ea7799d857fc6
GitOrigin-RevId: 0df8b0951163464d2a18971ebfd7d55b1efbe453
Change-Id: I38ecbcfa652d8faf4ef35a12ad67c18c3ae6240f