[roll] Roll fuchsia [suspend_test] Use lid to trigger resume.
The lid driver sets itself as a wake device in its implementation of the
suspend hook, so using the lid to trigger resume tests that hook as well
as the lid as a wake device. We use the lid instead of the power button
for this test because using the lid to suspend and resume will likely be
the most common workflow for using suspend.

Also add more traces during suspend and resume so the logs are more
helpful when the test fails.

Original-Bug: 107797

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748444
Original-Revision: ffc167442920eb9f02000ad9403772404e6e8ff2
GitOrigin-RevId: 536038cf2d588ce16f7c27f189713a27f7925001
Change-Id: I0cbc580c305b3e6f2ca4553f41d9ee6d0b99d232