[roll] Roll fuchsia [fxfs] Make sure we zero the tail of blocks on reads
We need to ensure that the tail of blocks is always zeroed on reads.  We
also need to ensure that the tail of blocks is zeroed when extending a
file.

The code to zero the tail on writes has been removed on the basis that
it won't be adding *much* benefit since Fxfs and the underlying storage
stack is COW, so the data will likely be present somewhere anyway.

Original-Fixed: 88676
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744827
Original-Revision: 3c21f38ee1c58ea4481e5b92e13318fd9c049044
GitOrigin-RevId: 96f313741ec014c8838129b04c8975b29d03bd1e
Change-Id: Ib7c759334c990a6ea4bba7244cd4e7e43322be01