[roll] Roll fuchsia [build][bazel] Record workspace events log.
This CL changes the generated `bazel` launcher script and .bazelrc
file to ensure that a log of workspace rule events is recorded each
time it is invoked.

This file is a binary that must be parsed with the Bazel Workspace
Log Parser program [1] to generate a text representation, that can
be used to detect undesirable download() and download_extract()
events triggered by workspace rules.

[1] https://github.com/bazelbuild/bazel/blob/master/src/tools/workspacelog/README.md

Original-Bug: 113230
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746123
Original-Revision: aaaf75567dde38807bf7d6dceef42e3f1cda8b2e
GitOrigin-RevId: 24872d52eca6b2aacf9c7ec33be1add36b573837
Change-Id: I209928044a641c97cefc68dc335b7cc0aeec0ab5