[roll] Roll fuchsia [netconform] Fix emulator deps
The GN Go build rules appear not to strictly enforce that dependencies
appear on the go_library target vs. the go_test target, so some
dependencies that were in the latter when they should have been in the
former resulted in an out-of-tree build breakage.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748650
Original-Revision: 7699eedfde7f19c2507b7ea0927560ed0f812c11
GitOrigin-RevId: a9b873f3ce79b8eb39852ad76c14eb4b3c8d499e
Change-Id: I08f6aac41c6fcf303557a29fc4962ec260bf881e