[roll] Roll fuchsia [component] Introduce `component_incoming` lib
This CL copies over the contents of `//sdk/lib/sys/component/cpp`
that pertains to FIDL client code into its own separate library.
The goal is to ultimately split up the library into an `incoming`
portion and an `outgoing` one. The motivation for this split is
that the `component::OutgoingDirectory` has a runtime dependency
on the `libsvc.so`. This is problematic for two reasons:

1: The shared library `libsvc.so` isn't always available to consumers
of the library. See the attached bug for a build breakage when trying
to build `libsyslog.so` which had a runtime dep on `libsvc.so`.

2: Not all components need the `OutgoingDirectory` portion of the
library, and its unnecessary to package a shared library that won't
be used.

Subsequent CL(s) will migrate all current users of `//sdk/lib/sys/component/cpp`
to this library.

Original-Bug: 113114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749342
Original-Revision: f4ae4bedc3f4446ac56673f7f3963d14b80200ba
GitOrigin-RevId: ebf113cda25c8bd338f057b08cb1bbcc5ac1dacd
Change-Id: I3840858c4ab768d599f79ec9fd790aa65aef13b9