[roll] Roll fuchsia [debugger] Move EvalContext reference into parser.
For local variable handling GetConcreteType() needs to be called from
the parser. Previously the parser and surrounding context information
was connected only via a "FindNameCallback" which is provided by the
EvalContext to the parser.

This replaces this callback when the full EvalContext. This is not
currently used so there should be no behavior change.

Remove the find name callback from the EvalContext, it is no longer
needed.

Allow ParsedIdentifiers to be used in a set/map by implementing
operator<.

Move the built-in type integration from the EvalContextImpl to the
FindName system. This cleans up some uses in tests which previously
didn't get, e.g. "int" types in C.

Hook up a FindName wrapper in EvalContext which automatically gets the
correct context. This also provides a point to inject mocks without
having to inject symbols into the full symbol system (previously this
was done via custom "find name callback". Adds this mocking capability
to MockEvalContext and use it in the parser unit tests.

Changes the parser test to pass the context by default (before
individual calls had to opt-into this by passing the name callback).
Most calls want this context and not having this has tripped me up while
writing tests since things like "int" won't even resolve in that case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748683
Original-Revision: ad67fe3fbc70a10af99614971559c75bba99e98a
GitOrigin-RevId: 8f53a5902c26649452315abc0c3a53ea80b2e0c4
Change-Id: Ic866fbcf0af989c4c74d693828690dcfef65ee9c