[roll] Roll fuchsia [bt][a2dp] Mark tracing as optional
This is a reland of commit 47341d71947f1a9fccc2732eb28639a99cc5e8d9
with the following changes:
 - Use tracing in source instead of sink in loopback as e2e test detects
   the packet sent tracing to verify data is sent.
 - Added comments to bt-a2dp-loopback to indicate this need.

fuchsia.tracing.provider.Registry is not available on all builds, and we
can continue to operate without it, so mark it as an optional protocol
in case tracing is unavailable.

Reformat optional protocols for brevity.

Original-Bug: b/248584553
Test: updated bt-a2dp-loopback to start source without registry
      run e2e tests to confirm they still can detect packets
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748944
Original-Revision: 3fc117252dd464a30b6f1150d6ccbab8c43008ec
GitOrigin-RevId: 1f7cdcf9f01df766416ab487e99049cbe3e561c6
Change-Id: I6b07ac34045e6f82730f181ee2f22f801816005f