[roll] Roll fuchsia [ui-input] Add _usecs to pointerinjector inspect values
Currently the values reported to inspect by the pointerinjector are
in nsecs, but that is not made clear anywhere. nsecs are also too
granular to really be helpful.
This CL changes the scale to usecs and adds a _usecs denominator to
the inspect field name.

Original-Fixed: 113413
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749489
Original-Revision: a1787e96a265e10e2392e375bdc690128fbccbef
GitOrigin-RevId: db2d9f9dcb5baafa2d9934adba9860bfb28e7fb8
Change-Id: Ib1455c2930250eb51f13a78ac688a01aff0d7453