[roll] Roll fuchsia [fint] Improve no-op check error messages
Currently if the ninja no-op check fails, the error message that's
propagated to the build results page will say only "Ninja build did not
converge to no-op", and users must dig into the ninja explain stderr to
figure out which target is causing the problem.

The first line of the explain stderr is usually enough to explain the
failure, so we should include that in the error message.

This will also help with writing queries to analyze the rates of
different types of no-op failures.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748771
Original-Revision: d1214f088ed5eea40b2c29fa3c16bb6bfc02be04
GitOrigin-RevId: 9df3fa05445cb758df5fb974538d56c2e1435d8f
Change-Id: I13d0fcebffa040015752cd12d336d7b68717019a