[roll] Roll fuchsia [line_input] Don't flush input.
Previously when we changed terminal modes we did a flush. But the kind
of flush that was used discarded all buffered input in stdin. This
creates a race with user typing and has the effect of randomly
discarding input.

Change to only flushing the output on state changes. This is the
important part because that can get buffered and the output formatting
depends on the terminal mode being correct.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749604
Original-Revision: eba8cf6bd7616ea433efcda4632d1a941bf751ad
GitOrigin-RevId: 1ca66f402971aeed8c64354b1676bfb315df3895
Change-Id: I8556c294e8b3771fc63c90a5a1aad83b2db7978d