[roll] Roll fuchsia [bt][socket] Fix SocketChannelRelay::OnSocketWritable() crash
Fix SocketChannelRelay crash caused by a failing
`!socket_write_queue_.empty()` assert in
SocketChannelRelay::OnSocketWritable(). This should not be an assert
because there is an edge case where the write queue may be emptied
before OnSocketWritable() is called. In OnChannelDataReceived(), if the
first packet in the write queue gets dropped and the subsequent smaller
packets all fit into the socket, the write queue will be emptied.

Cancel the socket writable wait in OnChannelDataReceived() if the packet
its threshold corresponds to is dropped.

Original-Fixed: 77407
Test: fx test bt-host-socket-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749182
Original-Revision: b2997b71162851daac21892714070b82fd26ac92
GitOrigin-RevId: 388607479213188e12e0a5f387038c8e790ce04c
Change-Id: Ia4a2094e031b5b2455d798fdb58b9f24b5584d1d