[roll] Roll fuchsia [scenic] Remove SCENIC_IGNORE_VSYNC
Deletes the SCENIC_IGNORE_VSYNC build-time argument.
As far as I know this build argument has been unused for years.
If this is a feature we want it should probably by a runtime switch,
or at the very least set as part of a config.

Original-Fixed: 24544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749642
Original-Revision: 657469fa00e9eb617a4018950c8c9c7830770b9f
GitOrigin-RevId: 3ea1fde5e2a39f227ebefd17efe2a28d1e6373b5
Change-Id: Ib27f93ad38eaf679265a74acb4060c984016e00e