[roll] Roll fuchsia [fdf] Update device group primary node logic
Update device groups so that they receive the
primary node from the composite driver.

Cleaned up MatchedDeviceGroupInfo by removing
the node values in its MatchedCompositeInfo
data.

The DFv1 and DFv2 implementation are updated to
reflect that. In DFv1, the primrary node index
is passed to CompositeDevice. In DFv2, the Node
class is modified so it can handle a primary
node index.

Original-Bug: 108360
Tests: Updated existing tests and added a
       device_group_v2_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747402
Original-Revision: 210960fdd6120c33665813587c29284315feabd7
GitOrigin-RevId: e3193e06a1baf9a19aa18cdd6f719dfd851877d8
Change-Id: I942f5ff96e9791c0d5d0bc2271eccff3cf948517