[roll] Roll fuchsia [inspector][gwp-asan] Ignore OOM errors.
GWP-ASan uses page faults to detect its own exceptions. However, if a
page fault occurs because the system ran out memory when trying to
access a gwp-asan tracked variable, the runtime will report it as a
gwp-asan error.

The error code for page faults can be found in the synth_code inside the
exception report's context.

This commit checks the error code of the page fault for out of memory
errors (`ZX_ERR_NO_MEMORY`), and if so, ignores the exception.

A test was added to simulate this code path. It is not possible to force
the system to generate an out-of-memory page fault via regular ways, so
instead we just change the synth_code of the exception report manually
for the test.

Original-Fixed: 113375
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748943
Original-Revision: d8add68998552cec98efaadf4170e3eafe3d6f98
GitOrigin-RevId: 3d78cace63062d65d3ad80a6ff9c208fa3b804d8
Change-Id: Id92a9fc1594eeb716fa78f46b187629660487026