[roll] Roll third_party/ffmpeg [third_party/ffmpeg] reference config dir more explicitly
The Fuchsia ffmpeg repo is integrated at
//builds/secondary/third_party/ffmpeg so that its build files are
mirrored in //third_party/ffmpeg. Unfortunately, the config files
in this repo are not mirrored. Consequently, the BUILD.gn file has to
be specific about the location of the config files so the compiler
can find them. This CL changes the `ffmpeg_config_root` variable to
point to the config files in //build/secondary. A comment has been
added explaining the situation and what to do if, in the future, we
are able to move this to //third_party/ffmpeg

Test: build issue
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/ffmpeg/+/748889
Original-Revision: 6d263d0900b7678ec74e1a8a78d9b99cc438870a
GitOrigin-RevId: db5b3c806b75c55f4572f75046033b013b437755
Change-Id: Ia7695d17ab8668e6061df35fd68a3edc9ec61247