[roll] Roll fuchsia [audio][test] Async variant of TestServerAndClient
Split TestServerAndClient into sync and async variants.

The async variant is currently hardcoded to use natural types;
the sync variant is currently hardcoded to use wire types. We
will generalize this in a subsequent CL, so that this can be
specified with an additional template parameter.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742093
Original-Revision: 325f8aae4965d777bbbc461c66c87c9652d134f1
GitOrigin-RevId: e16812c539de7f58e64f0f68be7c220c80d346af
Change-Id: I5b540ce9042d26b573981b4a12fadd414ba3c6a1