[roll] Roll fuchsia [netstack] Hold IntegralStatCounterMap by pointer
This fixes a data race caused by copying a tcpip.IntegralStatCounterMap during
reflective traversal. The reflective field access inadvertenly produces
a copy of the struct while trying to extract a pointer to the original
struct. Copying the struct copies both its fields: the
map[uint64]*StatCounter and the RWMutex that protects it. The new
RWMutex is distinct from the original, but because the map is a pointer
type, both the original and copied struct's map fields point to the same
underlying data. That means that their respective locks no longer
provide a guarantee of exclusive write/shared read access.

In the DHCP client struct, make the child tcpip.IntegralStatCounterMap
fields pointers instead of holding the structs inline. This ensures that
when the reflective traversal of counters is performed, it does not ever
produce a copy of an IntegralStatCounterMap.

Original-Fixed: 113437
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749605
Original-Revision: f957d471b6212d8d39aa3c2de87a46c64fef5141
GitOrigin-RevId: ae9564fca92a67572ecc22a27573974b9d7b31eb
Change-Id: I27e736aa11cbf2064ebe9309ff60c69ea23fe2b7