[roll] Roll fuchsia [build] Fix toolchain redirect
This group is intended to redirect test packages targets included in a
non-default toolchain to the corresponding target in the default
toolchain, but it was previously broken because it tried to use a
relative directory name rather than a relative target name.

I think this group will only ever be included when there's a bug in a GN
file that results in a target being included in the wrong toolchain.
That will now result in an `assert(is_fuchsia)` failing somewhere, which
is more helpful than a red herring "Unable to load <path>/BUILD.gn"
failure, which is what would happen before this fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750352
Original-Revision: 967ebdcb2f78fa1dcb14f0db9c75a32cfdf86737
GitOrigin-RevId: d8e0a5c77c0bc8f771f565424cb36019228386f4
Change-Id: I37dd151dc1d0ec419dfb5f6d7d2995665a3842a1