[roll] Roll fuchsia [kernel][vm] Fix pin_count assert
When transitioning from AwaitingClean to Clean, we cannot assume that
pin_count will be zero. It is possible for a page to be pinned for read
after having been marked AwaitingClean, in which case it should be valid
to mark the page Clean while pinned. It is also possible for pages to be
in an intermediate pinned state as a write-pin will only attempt to
dirty pages after having pinned them, so a racing writeback should be
able to clean pages if the dirty request has not been resolved yet.

We can only assert that the pin count is zero when transitioning from
Dirty to AwaitingClean, since the writeback code explicitly forbids this
transition.

Fix the incorrect assert on Clean transition and add tests that would
have tripped it.

Test: core-pager-writeback
Original-Bug: 106863
Original-Bug: 63989
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749615
Original-Revision: 08fb294814218e17e6347d38e3ed5a9dcac60c44
GitOrigin-RevId: 469a15eb1bf025edf6ef913b0329fb07e74d30c4
Change-Id: I938853af4282ffa9a7afab9f32b4157c806d4c3b