[roll] Roll fuchsia [intel-i915-tgl] DP: Always train link on display initialization.
On Tiger Lake, now that we unconditionally reset and reinitialize
all the display devices, it is possible that a link is already
trained on display initialization. In that case, the driver should
not simply treat it as "trained" but should still do the initialization
procedure again.

Note that this CL also changes the behavior of Kaby Lake on display
re-initialization though we didn't see any side effects.

Test: HDMI port on Intel NUC11PAHi5
      Chromebook (Atlas) with bind rule changed

Original-Bug: 109368
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744873
Original-Revision: 87d95211d3e9c451677c3106a90b2d3759ba5999
GitOrigin-RevId: 9de200bec33815c104a1af03a2526cf3d16e5186
Change-Id: If12ffcf987c8f38c5c60a193b951fa703b46ca0c