[roll] Roll fuchsia [fdf] Add device_controller connector
Add an fs::Service as a child of each device that's labelled
'device_controller'. Connecting to this service connects you directly to
fuchsia.device.Controller without any FIDL multiplexing.

These 'device_controller' files are added in /dev/class/ and in the
topological paths.

The DriverHost must now check on each Controller call if the device is
unbound. Normally this was handled in HandleFsSpecificMessage.

The test checks that this works in DFv1 and DFv2.

TEST: device_controller_fidl

Original-Bug: 112484

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744403
Original-Revision: 7eb6fe9aeea2b0d688864932a6cdd10cd163b0c4
GitOrigin-RevId: c4b6df91a28c7aa2eca5339daed46e79ee004502
Change-Id: I9ffd2684b3d7c144bcf1f472f9ff1cdea5be85b6