[roll] Roll fuchsia [bt][init] Use core-env for the #bluetooth-core realm
Currently, the bluetooth-core realm is not product variable. It is
defined in the base `core.cml` manifest.
`bt-rfcomm` and `bt-fastpair-provider` are product dependent children of
`bt-init`. To avoid churn during development, both packages are included
in the cached set of packages in eng builds. These two packages need the
full-resolver so that they can be re-resolved and updated ephemerally.

Use core-env for the BT realm so that the child components can be
resolved.

Original-Bug: 108241
Test: manifest changes only, fx build, fx ota, verify
bt-fastpair-provider is resolved correctly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750410
Original-Revision: c8e06c1e47f8b00ce746f71498266bfba28b0845
GitOrigin-RevId: fddb69734388b1e13294f95963cad9a4586fc9b0
Change-Id: I452d8166d12e158c5023c92a449a353f0853e69c