[roll] Roll fuchsia [virtualization] Post error when VM fails to launch.
If the VM fails to launch, vsh will not report any feedback because the
error is not sent over FIDL. This updates termina_guest_manager to post
and error to the client in this situation.

We also add a new error for failing to allocate a disk for a guest since
this is a common failure mode.

Original-Bug: 113232
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748747
Original-Revision: 251dd9880f22b9962722c221b551cdda513c959e
GitOrigin-RevId: 6a59d99735d850bddfc9437a4b2d146299a9c871
Change-Id: I66e0694d22745b0ca88db512c9d9ccb9c36c5151