[roll] Roll fuchsia [fxfs] Fix cause of test timeout
futures::lock::Mutex is particularly unfair and it's easy to cause
starvation.  LockManager locks are also not guaranteed to be fair but
when locks are unlocked, *all* wakers are woken up and so who gets the
lock next is dependent on the executor, which is good enough to avoid
the starvation issue we were seeing with the test_unlink_with_open_race
test.

Original-Fixed: 113620
Original-Bug: 102659
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751082
Original-Revision: e336b2d9f222b0520dcd909e203067bc6bbc129a
GitOrigin-RevId: 95ae72ae96a5a4b99234a32f3b203da57dbe7b43
Change-Id: I8b0145784b59918a4c481fcea123ed7767a38e88