[roll] Roll fuchsia [fdf] Support DdkOpen and DdkClose in compat shim
Support the ability for DFv1 drivers to listen to when their associated
devfs node gets opened and closed when DFv2 is enabled and the driver is
running in a compat shim.

Properly close connections to vnode associated with instance devices
when the device is destructed in the DFv2 compat shim.

Don't modify DFv2 compat device's name when creating an outgoing name
for the device.

Original-Bug: 102234
Original-Bug: 112361

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748783
Original-Revision: c3ceb7931da673f5d189df0eb4c6ec451d99daf7
GitOrigin-RevId: 93872190f8e61d2e4842ea1fafda6bb6bea28f8b
Change-Id: Ia535b321b7840d2d613d6e0a8ad2bda7f51e43c7