[roll] Roll fuchsia [forensics] Set signature for no memory page fault crashes
Programs page fault with a ZX_ERR_NO_MEMORY status when the system is
unable to allocate memory for them. Capture this reason in the crash
signature in order to prevent the signature from being based on the
stack trace, which is useless because the code is not fault for the
crash.

Original-Bug: 113439

TESTED=Not able to b/c this code path requires a real exception to be
generated and we don't have a way to create a ZX_ERR_NO_MEMORY page
fault on demand

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751342
Original-Revision: b017988ffe68081574c69da603062d46b045b079
GitOrigin-RevId: 63923a171c0d726a91667164b0d063f2cb2cc267
Change-Id: I4ef7c1e98027b9a421514dfd2e23698aeae819e2