[roll] Roll fuchsia [zircon] Generate <zircon/errors.h>
The current <zircon/errors.h> is meant to be consumed in both C and
assembly and so does not feature any C-style casts. To decouple any ABI
changes, we preserve this by using the zither asm backend for generating
the file for now.

Original-Bug: 51002
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738606
Original-Revision: df0debdd26a0048fe8be17bf5622cc856222672b
GitOrigin-RevId: 808b7aff9af92fea65fc460b1890017c44eeadb3
Change-Id: Id6644cbadccbac169013f54797fdc1fc80edb9b8