[roll] Roll fuchsia [xhci] Fix ERDP pointer read
This patch fixes an edge case where we might incorrectly read an event
TRB that HW hasn't handed control of over to us yet.

This occurs when we're reading a new segment for the first time, and so
validity checks should be done by completion code instead of cycle bit.
The buggy sequence is:
1. AdvanceErdp increments ERDP and returns a Control struct with its
   cycle bit set to !ccs_.
2. In HandleIRQ the cycle bit is checked and since it doesn't match ccs_
   we return.
3. We get a spurious interrupt.
4. We read from the ERDP, skipping the completion code check and not
   modifying the cycle bit. We enter the loop based on an invalid cycle
   bit check.

It's not clear whether this is benign (e.g. if we never get spurious
interrupts it's fine) but this change is at least more correct.

Original-Bug: 107934
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747289
Original-Revision: a6df301917ec9d413ff109a3243f3903bb137533
GitOrigin-RevId: 529bf75630a5dbd7f1843d41f4d319909117b1b5
Change-Id: I1ce827a370b4b983b2efd9287bbfe48e4c96e248