[roll] Roll fuchsia [netstack3] Update multicast membership w/ lock
When a datagram socket requests to join or leave a multicast group,
perform the operation while still holding the socket state lock. This
ensures that join and leave operations are atomic and moves us closer to
the goal of removing the lock protecting Ctx in the bindings.

Original-Bug: 108008
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750500
Original-Revision: 6dc0a87e8b531fc3aced6c8f965aee535d01ecc3
GitOrigin-RevId: b39265a87f79376136669ec6e68c858069269553
Change-Id: I47020e6eaccdae382c58cce66087cfe707533604