[roll] Roll cobalt Fix the source generation of name maps for Java.
There were a couple of problems:
- use of primitive type in Map definition
- extra "," at the end of the last Map entry line

I also changed the type of the event codes to Integer, to match the
use in the enum and make it easier to use with existing code that
expects integers. And I switched to a static initializer block, as
Map.ofEntries wasn't added until Java 9.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/751027
Original-Revision: 7d28bb713cddb7289037a7839a1d63abec1ed034
GitOrigin-RevId: 12854fc37d299e181d6ee716bf72715948c6c4e3
Change-Id: I14130557b343806c7836c0cf9f263ae09a15f4be