[roll] Roll fuchsia [perfetto-bridge] Added an integration test
This test set up the topology

         test-harness
               |
               |
          trace_manager
           /        \
perfetto-bridge---test_producer

Then starts and stops tracing and checks if perfetto data exists in the
trace buffer.

It does not yet emit perfetto data itself which will come in a follow up
patch that implements perfetto::Producer.

It does check that it's able to correctly make a connection and that
it's able to see the data that perfetto emits on startup.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735074
Original-Revision: ff2f63009d5ce31eeb8056ed5bc16d3263174884
GitOrigin-RevId: 63ee81df08ec1956a32c8d07c954f5e8e3597a2d
Change-Id: If5f905cd32d6343f91c885a650a028b4a59c1f39