[roll] Roll fuchsia [fidl][docs] Update FIDL design history
One of the "current status" categories, "Rejected; Superseded", was
removed from the list. I originally wanted to add "Withdrawn;
Superseded", but realized that this was going to create a lot of
needless combinatorial complexity for readers. Given that this column is
already a bit free form, I removed this line and figured users will
figure it out on their own.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752043
Original-Revision: 5829b3d8e449e74ff97750eca01560a17639e38c
GitOrigin-RevId: 169cbcdb16c66076da1cb8258e04a9d7b44ad39f
Change-Id: I049a505a2ecf547169584d98d86010cfd2576d24