[roll] Roll fuchsia [mdns] Use alternate host name for configured services
This CL adds an item to the configuration format that allows
'alternate services' to be specified. Alternate services are
identified as such by type and are published under an alternate
host name. The alternate host name is just the hexadecimal digits
from the host name, capitalized. So the alternate host name for
"fuchsia-1234-5678-9abc" is "123456789ABC".

The config file entry looks like this:

    "alt_services": [ "_foosvc._tcp.", "_barsvc._udp." ]

This tells the mdns service to publish "_foosvc._tcp." and
"_barsvc._udp." services under the alternate host name.

If any alternate service names are specified in the configuration,
the mdns service will respond to resolution requests for the
alternate host name, otherwise it will not. The service never
performs a conflict probefor the alternate host name.

Test: added/enhanced unit tests to prevent regression

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751404
Original-Revision: d2f7dbcbd8cceb94233839612a6a0ede9051b98c
GitOrigin-RevId: 6e13838d1cbf55976d564735723d46dad4ab1357
Change-Id: I87e8845ebec6a3cd3327d75d19a119fc13d9f20d