[roll] Roll fuchsia [emu] Add one-liner about network mode to 'start'
The attached bug describes how a user can be confused by the default
behavior of the --net flag in `ffx emu start`, leading to potential
support tickets. This CL adds some visibility around the resolution of
the "Auto" value by printing a single line to the terminal output after
resolution is complete. This line contains the mode selected, and a link
to a fuchsia.dev page with more information about networking modes. This
message is only printed when the "--net auto" flag is being resolved,
not when the networking mode is explicitly chosen by the user.

Original-Bug: 113058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750372
Original-Revision: 50e20c5f0692e758b0055ce9ac195deb550bf683
GitOrigin-RevId: 7e1561601efb0bbf9403acb9e989e2bbbb6bf24e
Change-Id: I5d741a2f194c50399f73b29ed9a8ee51440dfe9a