[roll] Roll fuchsia [wlandevicemonitor] Remove references to "v2" feature
The "v2" feature was added as a temporary feature to aid migration to
CFv2. That migration is complete, and the feature is no longer needed.

Many of the comments in the wlandevicemonitor/BUILD.gn file are also
removed. The comments were more valuable when we had duplicate build
targets for v1 and v2 components. Since the wlandevicemonitor build
targets are no longer unique in this way among all the other wlan
component build targets, the comments are redundant.

Original-Bug: 78327, 78050
Test: None. Removing dead code.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751205
Original-Revision: 271e1293e4aee8636bd012e2dd6de301f7fcddaa
GitOrigin-RevId: 439806fb5897f7b094ac1bca5e31107e5e92c503
Change-Id: I2eb3eb4dda84dd67d3b858aa8dc9a0738105494f