[roll] Roll fuchsia [netstack3] TCP cubic congestion control
This commit implements RFC 8312 for the TCP state machine in Core: it
implements the slow start and the congestion avoidance algorithm. Fast
convergence, fast retransmission and fast recovery are not in this
commit and will be addressed in a future commit. One difference worth
noting is that we count bytes instead of segments for our congestion
window as opposed to the RFC.

Original-Bug: 111778
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747062
Original-Revision: 6f269626360f4ea6b20587b5580cb1fa17ad4b0e
GitOrigin-RevId: ba3879aa33e55f1cd9dc589770d591bfe02979df
Change-Id: I2d28d5dda1dcd2be35b142a1c8206b9157b6509e