[roll] Roll fuchsia [wlan][sme] Replace ScanTransaction with return from Scan
SME now returns a full scan result to the caller instead of a protocol
the caller must stream results from. The direct return is enabled by
using a new FIDL attribute to enable sending messages of unbounded
size. This new approach has a negligible impact on latency and is
obviously much more ergonomic. The latency of the returned results
is can almost be viewed as irrelevant because the scan itself requires
seconds to complete, and latency is on the order microseconds.

The ScanError type is removed. The additional message string in that
type was never used in a meaningful way, and its removal makes the
Scan() method more ergnomic for Rust.

This is an initial implementation of the FIDL large messages RFC
available at fxrev.dev/695625.

Test: Existing tests modified. Tests added for large scan results.
Original-Bug: 88648, 107777

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705311
Original-Revision: 9d5c93b19ea1ee3fa9a4696dbcf75b615785ddcd
GitOrigin-RevId: 462785738aa386663a72736fbf56902b3552f49e
Change-Id: I146d1fd5d9002c81aaa253c6bd6167b59ab9bebe