[roll] Roll fuchsia [build][bazel] Add local rules_python repository.
This repository is required by SDK rules to perform assembly,
so we need a local version checked out in the Fuchsia tree to prevent
unexpected download.

This has already been approved by OSRB [1] but there is still no mirror
on fuchsia.googlesource.com for it [2]. Since this is blocking deployment
of Bazel on CI/CQ, create a local copy until the mirror is setup
properly and a Jiri manifest entry is added to it (which will point
to //third_party/bazel_rules_python instead to allow for a soft migration).

Note that these files have been reformated by `fx format-code` because
this tool will always pass over all committed files, even if they are
listed in the .gitgnore. This problem will not happen when using a
proper Jiri manifest entry where files come out of fuchsia.git.

+ Add an exceptions for //build/bazel/local_repositories in
  tools/check-licenses/directory/_config.json, otherwise the tool
  complains about missing Fuchsia copyright headers.

[1] https://fxbug.dev/113227
[2] https://fxbug.dev/113330

Original-Bug: 112781
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748587
Original-Revision: fb5306e16857128870bb3f0030b5ef575667d862
GitOrigin-RevId: 15527796251128c5834db63af27f9b4d2c44b87e
Change-Id: I58d15e0c755c5fcf15f41469f5b425611e2ca9c4