[roll] Roll fuchsia [bt][hfp] Fix crash formatting codecs as strings
Previously, if the AG recevied a codec ID from a peer that was not one
specified in the HFP spec (either 0x1 0x2), it would panic attempting to
format it as a string.  This means that noncompliant peers can crash
HFP. We are seeing a few of these a day in the field.

This change adds support for unknown codecs to the string formater for
code IDs and adds a test for this.

Test: fx test fuchsia-pkg://fuchsia.com/bt-hfp-audio-gateway-tests#meta/test-component.cm
Original-Bug: fxb/113474
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752983
Original-Revision: a0f251613dccea8556d7ce908a4663ebdd04bfc1
GitOrigin-RevId: 80bc417824f75e251640ce444fcbe46acdbd255c
Change-Id: I4fe993be70e3187ffe31b3eb3e80a815ab7f4c57