[roll] Roll fuchsia [vaapi][fuzz] Workaround one buffer collection
Current CodecImpl only supports one set of buffers allocated at a time.
This is a problem for the TiledSurfaceManager since reference frames
that are allocated prior to the buffer reallocation will be destroyed
and are unable to be used after. Added some temporary workaround logic
that detects if a non-keyframe requires a buffer reallocation and
ensures that the current implementation of surface manager supports it.

Original-Bug: fxbug.dev/110529
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753042
Original-Revision: 98943092a5c695a2b91c54f8a58baf21e3a7716b
GitOrigin-RevId: 9c88d8e7e3c79017ce43b32bda71e08af2b2facb
Change-Id: Ie678ec853bb28113e4510e661ec0f3670ce18d88