[roll] Roll fuchsia [virtualization] Batch completions in the net device
Calling into the netstack is synchronous and can be slow, which is
why batched completions are supported. Conversely, waiting to batch
completions can add latency.

A compromise is to still send completions as quickly as possible, but
accrue completions in a queue for batching. For example, the C++ thread
schedules a "complete batches" task, which may be scheduled behind
multiple other tasks. While this waits to run, the Rust thread can
continue to append things to complete to the queue.

Note: This doesn't wire this up to the device, that'll be done in a
second CL to simplify the review and iperf testing.

Original-Bug: 95485
Tested:
Unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751042
Original-Revision: 446ec29b808e1829cb1d71e124e4bd55b13e13ca
GitOrigin-RevId: f149a7afdff9674083a5216863f9370d7eac3fc6
Change-Id: I0ac9878911c1aa1b68018e0176b5c3b7960827fe