[roll] Roll fuchsia [component_manager][events] Fix leak in event source
Spawn the event stream server task in the subscriber's task scope,
instead of the `task_scope` passed to `open` (which in this case is the
global one because event streams route to above-root).

This ensures that when the subscriber is destroyed, the task is
cancelled.

Original-Bug: 110261
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752984
Original-Revision: ff8cd9443ccb32dc7474d5cafc16b0a1ad9f1c77
GitOrigin-RevId: dfb18fe856fc949bc54202178a7a47aaa06385b9
Change-Id: I532a2506da1a3f69a7024034348745ff1e759464