[roll] Roll fuchsia [basemgr] Fix basemgr lifecycle race.
basemgr was updated ~9 mos ago with a second implementation of process.lifecycle.Lifecycle. This implementation, which was bound to PA_LIFECYCLE handle, does not wait for BasemgrImpl's shutdown sequence to complete before stopping the main async loop.

The solution is to use the implementation built into BasemgrImpl, which calls an `on_shutdown` callback on completion of the shutdown sequence, which in turn stops the main loop.

Testing: added a unittest
Original-Bug: 112090
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749287
Original-Revision: 4a7dcb916c321aa932d403cda16487ebf8fbe336
GitOrigin-RevId: 010b68bb3ebe8ffb6e76f52d88ae8d9c50e97557
Change-Id: Ib7fd35184b904f5fba836d708416ebcb8359f8c3