[roll] Roll fuchsia [image-format] Support bytes_per_row with Intel tiling
Intel Gen12 GPUs require that tiled images are multiples of 4x1 tiles
for use with color compression (even if the color compression buffer is
private to the driver and not exposed externally through sysmem). To
enforce that in sysmem, allow specifying bytes_per_row on Intel tiled
formats.

The "row size" is the number of bytes of a row of a linearized
view of a texture.

Original-Bug: 109075
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748686
Original-Revision: d4279143b19d06138bd2148d3cd3810cb633441c
GitOrigin-RevId: 240d49867597471fdd61b2fedce4ff4c361fd77a
Change-Id: I50010abc5cfa2186c77c7f460bd7f2ea2ed71362