[roll] Roll fuchsia [fhcp] Fail the touchpad test if erroneous input is detected
The test now assumes that the manual tester is infallible and that any
unexpected input (for example, a touch in the wrong quadrant) is the
result of a bug in the touchpad driver.

This allows for the simplification of some of the logic of the test.

Original-Bug: b/255810966
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752782
Original-Revision: b017ddec53cd908acbba5bdbc1d8d57e4f9f4a44
GitOrigin-RevId: 976ecc917fdf127af77208d4d7dd050a67ab0905
Change-Id: I5f000c6c971a4401f8bb2864a8f1cb1f9fd52ffa