[roll] Roll fuchsia [pb] Move the MOS files outside the PB
The files intended for MOS (all_blobs, targets, images) must be uploaded
to GCS relative to the product bundle in the same way that it is
relative in the build directory. This is because these MOS files have
relative paths in them.

This CL addresses an issue where the images.json file was moved during
upload, causing the relative paths to be invalid. We now upload the MOS
files next to the product bundle just as they are uploaded to GCS.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753386
Original-Revision: e977642eb7a835c9e1e7e5f7c3ea2c1ba2666320
GitOrigin-RevId: 5493836b1ebcf7661b03dcbfd17a049a2dfeb01f
Change-Id: Ic0bfec56c2c7e1fbe86f760ff6b4ab982b8e3e3c