[roll] Roll fuchsia [usb-cdc-ecm] Check descriptor lengths
Check the length of the descriptor so that malformed descriptors don't
cause a buffer overflow. Checking the length from the header is good
enough here because the usb iterator will check the length field against
the size of the descriptor buffer for us.

Fix: 111684
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753567
Original-Revision: 3076d501c7d0c04b77546d043245d09b68863976
GitOrigin-RevId: e101b44a2fd514f4fa7ec023ccaa00aa7ec7344e
Change-Id: I1f26d6e637421015deb5de6e2f453ac6531a7830