[roll] Roll fuchsia [sdk] Fix depfile generation in cmc and use via _cmc_tool in the GN SDK
The cmc tool now emits a depfile for .cml files with no imports. This is
necessary, as otherwise the build system always determines that .cm
files must be rebuilt due to the absence of a depfile. This led to
unnecessary work during builds.

The cmc_compile template in the GN SDK properly emits a depfile, but the
_cmc_tool was failing to forward the `depfile` variable from its
invoker, so it wasn't actually being used in the build. This led to .cm
targets not being built when only included .cml files were modified.

Original-Bug: 114310
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753627
Original-Revision: 659b24e2308e17c965dc0ab83184b79f70386b04
GitOrigin-RevId: 6d3dd239760e7940c85f0eaaa3003a0303dc206b
Change-Id: If6cc4b8d83b4f4620e133ba1dd90f4a872c75594